% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthModelKeycloak.r
\name{AuthModelKeycloak}
\alias{AuthModelKeycloak}
\title{AuthModelKeycloak Class}
\description{
AuthModelKeycloak Class

AuthModelKeycloak Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{username}}{}

\item{\code{realm}}{}

\item{\code{endpoint}}{}

\item{\code{redirectURI}}{}

\item{\code{clientId}}{}

\item{\code{clientSecret}}{}

\item{\code{updatedAt}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AuthModelKeycloak$new()}}
\item \href{#method-toJSON}{\code{AuthModelKeycloak$toJSON()}}
\item \href{#method-fromJSON}{\code{AuthModelKeycloak$fromJSON()}}
\item \href{#method-toJSONString}{\code{AuthModelKeycloak$toJSONString()}}
\item \href{#method-fromJSONString}{\code{AuthModelKeycloak$fromJSONString()}}
\item \href{#method-clone}{\code{AuthModelKeycloak$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthModelKeycloak$new(
  username,
  realm,
  endpoint,
  redirectURI,
  clientId,
  clientSecret,
  updatedAt
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthModelKeycloak$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthModelKeycloak$fromJSON(AuthModelKeycloakJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthModelKeycloak$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthModelKeycloak$fromJSONString(AuthModelKeycloakJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthModelKeycloak$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
