% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthModelLdap.r
\name{AuthModelLdap}
\alias{AuthModelLdap}
\title{AuthModelLdap Class}
\description{
AuthModelLdap Class

AuthModelLdap Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{username}}{}

\item{\code{password}}{}

\item{\code{basedn}}{}

\item{\code{cacert}}{}

\item{\code{scope}}{}

\item{\code{userAttr}}{}

\item{\code{groupAttr}}{}

\item{\code{groupFilter}}{}

\item{\code{userFilter}}{}

\item{\code{binddn}}{}

\item{\code{endpoint}}{}

\item{\code{advanced}}{}

\item{\code{updatedAt}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AuthModelLdap$new()}}
\item \href{#method-toJSON}{\code{AuthModelLdap$toJSON()}}
\item \href{#method-fromJSON}{\code{AuthModelLdap$fromJSON()}}
\item \href{#method-toJSONString}{\code{AuthModelLdap$toJSONString()}}
\item \href{#method-fromJSONString}{\code{AuthModelLdap$fromJSONString()}}
\item \href{#method-clone}{\code{AuthModelLdap$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthModelLdap$new(
  username,
  password,
  basedn,
  cacert,
  scope,
  userAttr,
  groupAttr,
  groupFilter,
  userFilter,
  binddn,
  endpoint,
  advanced,
  updatedAt
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthModelLdap$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthModelLdap$fromJSON(AuthModelLdapJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthModelLdap$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthModelLdap$fromJSONString(AuthModelLdapJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthModelLdap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
