% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpuModel.r
\name{CpuModel}
\alias{CpuModel}
\title{CpuModel Class}
\description{
CpuModel Class

CpuModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{version}}{}

\item{\code{cpus}}{}

\item{\code{architecture}}{}

\item{\code{cpu_op_models}}{}

\item{\code{byte_order}}{}

\item{\code{threads_per_core}}{}

\item{\code{cores_per_socket}}{}

\item{\code{sockets}}{}

\item{\code{vendor}}{}

\item{\code{modelNumber}}{}

\item{\code{cores}}{}

\item{\code{modelName}}{}

\item{\code{frequency}}{}

\item{\code{frequency_max}}{}

\item{\code{frequency_min}}{}

\item{\code{virtualization}}{}

\item{\code{family}}{}

\item{\code{l1d_cache}}{}

\item{\code{l1i_cache}}{}

\item{\code{l2_cache}}{}

\item{\code{l3_cache}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CpuModel$new()}}
\item \href{#method-toJSON}{\code{CpuModel$toJSON()}}
\item \href{#method-fromJSON}{\code{CpuModel$fromJSON()}}
\item \href{#method-toJSONString}{\code{CpuModel$toJSONString()}}
\item \href{#method-fromJSONString}{\code{CpuModel$fromJSONString()}}
\item \href{#method-clone}{\code{CpuModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CpuModel$new(
  version,
  cpus,
  architecture,
  cpu_op_models,
  byte_order,
  threads_per_core,
  cores_per_socket,
  sockets,
  vendor,
  modelNumber,
  cores,
  modelName,
  frequency,
  frequency_max,
  frequency_min,
  virtualization,
  family,
  l1d_cache,
  l1i_cache,
  l2_cache,
  l3_cache
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CpuModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CpuModel$fromJSON(CpuModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CpuModel$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CpuModel$fromJSONString(CpuModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CpuModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
