% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatumModel.r
\name{DatumModel}
\alias{DatumModel}
\title{DatumModel Class}
\description{
DatumModel Class

DatumModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{version}}{}

\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{tags}}{}

\item{\code{class}}{}

\item{\code{dvs}}{}

\item{\code{source}}{}

\item{\code{url}}{}

\item{\code{remote}}{}

\item{\code{gitaccess}}{}

\item{\code{redshift}}{}

\item{\code{s3access}}{}

\item{\code{k8svolume}}{}

\item{\code{hostpath}}{}

\item{\code{nfsaccess}}{}

\item{\code{gcsaccess}}{}

\item{\code{generated}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatumModel$new()}}
\item \href{#method-toJSON}{\code{DatumModel$toJSON()}}
\item \href{#method-fromJSON}{\code{DatumModel$fromJSON()}}
\item \href{#method-toJSONString}{\code{DatumModel$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DatumModel$fromJSONString()}}
\item \href{#method-clone}{\code{DatumModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatumModel$new(
  version,
  name,
  description,
  tags,
  class,
  dvs,
  source,
  url,
  remote,
  gitaccess,
  redshift,
  s3access,
  k8svolume,
  hostpath,
  nfsaccess,
  gcsaccess,
  generated
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatumModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatumModel$fromJSON(DatumModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatumModel$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatumModel$fromJSONString(DatumModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatumModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
