% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatumModelGeneratedDetails.r
\name{DatumModelGeneratedDetails}
\alias{DatumModelGeneratedDetails}
\title{DatumModelGeneratedDetails Class}
\description{
DatumModelGeneratedDetails Class

DatumModelGeneratedDetails Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{git}}{}

\item{\code{model}}{}

\item{\code{tensorboard}}{}

\item{\code{source}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatumModelGeneratedDetails$new()}}
\item \href{#method-toJSON}{\code{DatumModelGeneratedDetails$toJSON()}}
\item \href{#method-fromJSON}{\code{DatumModelGeneratedDetails$fromJSON()}}
\item \href{#method-toJSONString}{\code{DatumModelGeneratedDetails$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DatumModelGeneratedDetails$fromJSONString()}}
\item \href{#method-clone}{\code{DatumModelGeneratedDetails$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatumModelGeneratedDetails$new(git, model, tensorboard, source)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatumModelGeneratedDetails$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatumModelGeneratedDetails$fromJSON(DatumModelGeneratedDetailsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatumModelGeneratedDetails$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatumModelGeneratedDetails$fromJSONString(DatumModelGeneratedDetailsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatumModelGeneratedDetails$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
