% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DfabApi.r
\name{DfabApi}
\alias{DfabApi}
\title{Dfab operations}
\description{
dkube_api.Dfab
}
\section{Methods}{

\describe{

cluster_status API to fetch the status of various clusters


groups_create_one_or_update API to create a new group if not present and update group with pools and users.


groups_delete_list API to delete a list of groups


groups_get_collection_all API to get all the groups.


groups_get_collection_one API to get all the groups.


nodes_get_all API to get collection of nodes.


pools_create_one_or_update API to create a new pool if not present and update group with pools and users.


pools_delete_list API to delete a list of pools


pools_get_collection_all API to get all the pools.


pools_get_collection_one API to get collection of single pool.

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DfabApi$new()}}
\item \href{#method-cluster_status}{\code{DfabApi$cluster_status()}}
\item \href{#method-groups_create_one_or_update}{\code{DfabApi$groups_create_one_or_update()}}
\item \href{#method-groups_delete_list}{\code{DfabApi$groups_delete_list()}}
\item \href{#method-groups_get_collection_all}{\code{DfabApi$groups_get_collection_all()}}
\item \href{#method-groups_get_collection_one}{\code{DfabApi$groups_get_collection_one()}}
\item \href{#method-nodes_get_all}{\code{DfabApi$nodes_get_all()}}
\item \href{#method-pools_create_one_or_update}{\code{DfabApi$pools_create_one_or_update()}}
\item \href{#method-pools_delete_list}{\code{DfabApi$pools_delete_list()}}
\item \href{#method-pools_get_collection_all}{\code{DfabApi$pools_get_collection_all()}}
\item \href{#method-pools_get_collection_one}{\code{DfabApi$pools_get_collection_one()}}
\item \href{#method-clone}{\code{DfabApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DfabApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cluster_status"></a>}}
\if{latex}{\out{\hypertarget{method-cluster_status}{}}}
\subsection{Method \code{cluster_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DfabApi$cluster_status(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-groups_create_one_or_update"></a>}}
\if{latex}{\out{\hypertarget{method-groups_create_one_or_update}{}}}
\subsection{Method \code{groups_create_one_or_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DfabApi$groups_create_one_or_update(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-groups_delete_list"></a>}}
\if{latex}{\out{\hypertarget{method-groups_delete_list}{}}}
\subsection{Method \code{groups_delete_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DfabApi$groups_delete_list(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-groups_get_collection_all"></a>}}
\if{latex}{\out{\hypertarget{method-groups_get_collection_all}{}}}
\subsection{Method \code{groups_get_collection_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DfabApi$groups_get_collection_all(notonboarded, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-groups_get_collection_one"></a>}}
\if{latex}{\out{\hypertarget{method-groups_get_collection_one}{}}}
\subsection{Method \code{groups_get_collection_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DfabApi$groups_get_collection_one(group, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodes_get_all"></a>}}
\if{latex}{\out{\hypertarget{method-nodes_get_all}{}}}
\subsection{Method \code{nodes_get_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DfabApi$nodes_get_all(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pools_create_one_or_update"></a>}}
\if{latex}{\out{\hypertarget{method-pools_create_one_or_update}{}}}
\subsection{Method \code{pools_create_one_or_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DfabApi$pools_create_one_or_update(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pools_delete_list"></a>}}
\if{latex}{\out{\hypertarget{method-pools_delete_list}{}}}
\subsection{Method \code{pools_delete_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DfabApi$pools_delete_list(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pools_get_collection_all"></a>}}
\if{latex}{\out{\hypertarget{method-pools_get_collection_all}{}}}
\subsection{Method \code{pools_get_collection_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DfabApi$pools_get_collection_all(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pools_get_collection_one"></a>}}
\if{latex}{\out{\hypertarget{method-pools_get_collection_one}{}}}
\subsection{Method \code{pools_get_collection_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DfabApi$pools_get_collection_one(pool, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DfabApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
