% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DkubeApi.r
\name{DkubeApi}
\alias{DkubeApi}
\title{Dkube operations}
\description{
dkube_api.Dkube
}
\section{Methods}{

\describe{

abort_migration API to terminate migration


add_cicd_image API to add new cicd image


attach_user_profile Attach profile to a user


create_project API to Create a Project


create_user_profile Create a user profile in dkube


datum_one_abort_download API to abort a download


datum_one_restart_download API to restart a failed datum download


datum_refresh_one API to refresh the contents of an existing downloaded datum


datums_add_one API to add a datum.


datums_archive_by_class API to archive  or un-archive all datums by class. Refer &#39;DatumClassType&#39; definition for possible values of class.


datums_delete_by_class API to delete all datums by class. Refer &#39;DatumClassType&#39; definition for possible values of class.


datums_deploy_one_model API to deploy a model


datums_get_by_class API to get all datums by class. Refer &#39;DatumClassType&#39; definition for possible values of class.


datums_get_metrics API to get metrics of a list of datum


datums_get_one_by_class API to get a datums of a class. Refer &#39;DatumClassType&#39; definition for possible values of class.


datums_get_one_version_lineage API to get lineage of a datum version


datums_get_one_version_metrics API to get metrics of a model


datums_publish_one_model API to publish a model


datums_release_one_model API to release a model


datums_testdeploy_one_model API to deploy a model in dev cluster


delete_cicd_images API to delete cicd images


delete_migration API to terminate migration


delete_model_catalog_item API to delete modelcatalog version item


delete_pipeline_experiment API to delete a pipeline experiment


delete_pipeline_experimentby_id API to delete a pipeline experiment by id


delete_pipeline_experiments_all API to delete all pipeline experiments


delete_pipeline_run API to delete a pipeline run


delete_pipeline_run_by_id API to delete a pipeline run by id


delete_pipeline_runs_all API to delete all pipeline runs


diag API to manage admin related operations.


dkubeinfo API to fetch all the information about dkube installation


dl_frameworks API to fetch all the frameworks


dlsupport API to fetch all the dl support by dkube


featureset_add_one API to create a fatureset


featureset_commit_version API to commit all output faturesets


featureset_copy_version API to copy featureset version to users directory


featureset_copy_version_abort API to copy featureset version to users directory


featureset_copy_version_status API to copy featureset version to users directory


featureset_delete API to delete a featureset


featureset_get API to get featureset configuration


featureset_get_one_version_lineage API to get lineage of a featureset version


featureset_list API to list faturestores


featureset_put API to update featureset configuration


featureset_update_version API to update fatureset version


featuresets_archive API to archive  or un-archive featuresets


featurespec_list API to list featurespecs for a spec featureset


g_p_us_capability API to get GPU&#39;s capability of a user, job, cluster, node.


get_all_cicd_images API to get all cicd images


get_all_datum_prev_cred API to get previous datum creation auth credentials of all users in dkube.


get_all_project_submissions API to Get all submissions of a project


get_all_projects API to Get All Projects


get_datum_prev_cred API to get previous datum creation auth credentials.


get_dvs_config API to list dkube version management system config.


get_job_groups API to get all job gropus.


get_migration_metadata API to export a single job.


get_one_project API to Get Single Project


get_one_run_lineage API to get lineage of a run


get_role API to get details of a required role.


get_session_key API to get session key(static jwt token)


get_single_user_profile Fetch single user profile


inference_monitoring API to update inference job.


job_get_collection_all_by_class API to get jobs collection for all jobs of a class.


job_group_add_one API to add a job group.


jobmigration API to migrate jobs


jobs_abort_one API to abort a single job.


jobs_add_one API to add a new job.


jobs_delete_one_by_uuid API to delete a job by uuid.


jobs_get_by_class API to get jobs of a user by class.


jobs_get_collection_all API to get jobs collection for all jobs.


jobs_get_collection_one API to get jobs collection for single job.


jobs_get_one_by_uuid API to get a job by uuid.


jobs_list_archive_by_class API to archive or un-archive jobs by class. Refer to &#39;JobClassType&#39; definition for possible values of class.


jobs_list_delete_by_class API to delete jobs by class. Refer to &#39;JobClassType&#39; definition for possible values of class.


jobs_start_one API to start an aborted job.


jsonform API to fetch json form for slurm job scheduling


list_user_profiles list all user profiles


login API to login into dkube. Unique API key is returned using which user can access dkube apis. API key access will be scoped based on the role of the user.


logout API to logoff from the dkube. The API key will be revoked and invalidated.


logs_get_one API to get logs of a job


migration_complete API to notify migration complete


migration_status API to send status of migration


migration_status_user API to send status of migration


model_deploy API to deploy model


notification API to post notifications to Dkube API server


projects_delete_list API to delete a list of projects


repo_update_one API to update a datum.


revoke_session_key API to renew the session key(static jwt token)


run_group_update API to update a run group.


source_migration_init API to initiate_migration


start_tensorboard API to start tensorboard for a given job.


stop_tensorboard API to stop tensorboard for a given job.


storage_volume_export Export the given artifacts as k8s volumes


storage_volume_reclaim Reclaims the volumes related to the unique ID


submit_project_metrics API to post submission for a project


target_migration_init API to initiate_migration


tokeninfo API to fetch all claims for a jwt token


trigger_runs_by_condition Triggers runs in dkube which satisfies the condition specified


update_datum_version_model Update version model info of a datum.


update_inference API to update inference job.


update_model_catalog API to update modelcatalog version item


update_one_project API to update single Project


update_run_one Update description of a run.


update_template_one API to update a template.


user_capability API to get capabilities of a user.


user_get_collection_one API to get user collection for single user.


users_get_collection_all API to get user collection for all users.


volumes_storage API to fetch information about storage.

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DkubeApi$new()}}
\item \href{#method-abort_migration}{\code{DkubeApi$abort_migration()}}
\item \href{#method-add_cicd_image}{\code{DkubeApi$add_cicd_image()}}
\item \href{#method-attach_user_profile}{\code{DkubeApi$attach_user_profile()}}
\item \href{#method-create_project}{\code{DkubeApi$create_project()}}
\item \href{#method-create_user_profile}{\code{DkubeApi$create_user_profile()}}
\item \href{#method-datum_one_abort_download}{\code{DkubeApi$datum_one_abort_download()}}
\item \href{#method-datum_one_restart_download}{\code{DkubeApi$datum_one_restart_download()}}
\item \href{#method-datum_refresh_one}{\code{DkubeApi$datum_refresh_one()}}
\item \href{#method-datums_add_one}{\code{DkubeApi$datums_add_one()}}
\item \href{#method-datums_archive_by_class}{\code{DkubeApi$datums_archive_by_class()}}
\item \href{#method-datums_delete_by_class}{\code{DkubeApi$datums_delete_by_class()}}
\item \href{#method-datums_deploy_one_model}{\code{DkubeApi$datums_deploy_one_model()}}
\item \href{#method-datums_get_by_class}{\code{DkubeApi$datums_get_by_class()}}
\item \href{#method-datums_get_metrics}{\code{DkubeApi$datums_get_metrics()}}
\item \href{#method-datums_get_one_by_class}{\code{DkubeApi$datums_get_one_by_class()}}
\item \href{#method-datums_get_one_version_lineage}{\code{DkubeApi$datums_get_one_version_lineage()}}
\item \href{#method-datums_get_one_version_metrics}{\code{DkubeApi$datums_get_one_version_metrics()}}
\item \href{#method-datums_publish_one_model}{\code{DkubeApi$datums_publish_one_model()}}
\item \href{#method-datums_release_one_model}{\code{DkubeApi$datums_release_one_model()}}
\item \href{#method-datums_testdeploy_one_model}{\code{DkubeApi$datums_testdeploy_one_model()}}
\item \href{#method-delete_cicd_images}{\code{DkubeApi$delete_cicd_images()}}
\item \href{#method-delete_migration}{\code{DkubeApi$delete_migration()}}
\item \href{#method-delete_model_catalog_item}{\code{DkubeApi$delete_model_catalog_item()}}
\item \href{#method-delete_pipeline_experiment}{\code{DkubeApi$delete_pipeline_experiment()}}
\item \href{#method-delete_pipeline_experimentby_id}{\code{DkubeApi$delete_pipeline_experimentby_id()}}
\item \href{#method-delete_pipeline_experiments_all}{\code{DkubeApi$delete_pipeline_experiments_all()}}
\item \href{#method-delete_pipeline_run}{\code{DkubeApi$delete_pipeline_run()}}
\item \href{#method-delete_pipeline_run_by_id}{\code{DkubeApi$delete_pipeline_run_by_id()}}
\item \href{#method-delete_pipeline_runs_all}{\code{DkubeApi$delete_pipeline_runs_all()}}
\item \href{#method-diag}{\code{DkubeApi$diag()}}
\item \href{#method-dkubeinfo}{\code{DkubeApi$dkubeinfo()}}
\item \href{#method-dl_frameworks}{\code{DkubeApi$dl_frameworks()}}
\item \href{#method-dlsupport}{\code{DkubeApi$dlsupport()}}
\item \href{#method-featureset_add_one}{\code{DkubeApi$featureset_add_one()}}
\item \href{#method-featureset_commit_version}{\code{DkubeApi$featureset_commit_version()}}
\item \href{#method-featureset_copy_version}{\code{DkubeApi$featureset_copy_version()}}
\item \href{#method-featureset_copy_version_abort}{\code{DkubeApi$featureset_copy_version_abort()}}
\item \href{#method-featureset_copy_version_status}{\code{DkubeApi$featureset_copy_version_status()}}
\item \href{#method-featureset_delete}{\code{DkubeApi$featureset_delete()}}
\item \href{#method-featureset_get}{\code{DkubeApi$featureset_get()}}
\item \href{#method-featureset_get_one_version_lineage}{\code{DkubeApi$featureset_get_one_version_lineage()}}
\item \href{#method-featureset_list}{\code{DkubeApi$featureset_list()}}
\item \href{#method-featureset_put}{\code{DkubeApi$featureset_put()}}
\item \href{#method-featureset_update_version}{\code{DkubeApi$featureset_update_version()}}
\item \href{#method-featuresets_archive}{\code{DkubeApi$featuresets_archive()}}
\item \href{#method-featurespec_list}{\code{DkubeApi$featurespec_list()}}
\item \href{#method-g_p_us_capability}{\code{DkubeApi$g_p_us_capability()}}
\item \href{#method-get_all_cicd_images}{\code{DkubeApi$get_all_cicd_images()}}
\item \href{#method-get_all_datum_prev_cred}{\code{DkubeApi$get_all_datum_prev_cred()}}
\item \href{#method-get_all_project_submissions}{\code{DkubeApi$get_all_project_submissions()}}
\item \href{#method-get_all_projects}{\code{DkubeApi$get_all_projects()}}
\item \href{#method-get_datum_prev_cred}{\code{DkubeApi$get_datum_prev_cred()}}
\item \href{#method-get_dvs_config}{\code{DkubeApi$get_dvs_config()}}
\item \href{#method-get_job_groups}{\code{DkubeApi$get_job_groups()}}
\item \href{#method-get_migration_metadata}{\code{DkubeApi$get_migration_metadata()}}
\item \href{#method-get_one_project}{\code{DkubeApi$get_one_project()}}
\item \href{#method-get_one_run_lineage}{\code{DkubeApi$get_one_run_lineage()}}
\item \href{#method-get_role}{\code{DkubeApi$get_role()}}
\item \href{#method-get_session_key}{\code{DkubeApi$get_session_key()}}
\item \href{#method-get_single_user_profile}{\code{DkubeApi$get_single_user_profile()}}
\item \href{#method-inference_monitoring}{\code{DkubeApi$inference_monitoring()}}
\item \href{#method-job_get_collection_all_by_class}{\code{DkubeApi$job_get_collection_all_by_class()}}
\item \href{#method-job_group_add_one}{\code{DkubeApi$job_group_add_one()}}
\item \href{#method-jobmigration}{\code{DkubeApi$jobmigration()}}
\item \href{#method-jobs_abort_one}{\code{DkubeApi$jobs_abort_one()}}
\item \href{#method-jobs_add_one}{\code{DkubeApi$jobs_add_one()}}
\item \href{#method-jobs_delete_one_by_uuid}{\code{DkubeApi$jobs_delete_one_by_uuid()}}
\item \href{#method-jobs_get_by_class}{\code{DkubeApi$jobs_get_by_class()}}
\item \href{#method-jobs_get_collection_all}{\code{DkubeApi$jobs_get_collection_all()}}
\item \href{#method-jobs_get_collection_one}{\code{DkubeApi$jobs_get_collection_one()}}
\item \href{#method-jobs_get_one_by_uuid}{\code{DkubeApi$jobs_get_one_by_uuid()}}
\item \href{#method-jobs_list_archive_by_class}{\code{DkubeApi$jobs_list_archive_by_class()}}
\item \href{#method-jobs_list_delete_by_class}{\code{DkubeApi$jobs_list_delete_by_class()}}
\item \href{#method-jobs_start_one}{\code{DkubeApi$jobs_start_one()}}
\item \href{#method-jsonform}{\code{DkubeApi$jsonform()}}
\item \href{#method-list_user_profiles}{\code{DkubeApi$list_user_profiles()}}
\item \href{#method-login}{\code{DkubeApi$login()}}
\item \href{#method-logout}{\code{DkubeApi$logout()}}
\item \href{#method-logs_get_one}{\code{DkubeApi$logs_get_one()}}
\item \href{#method-migration_complete}{\code{DkubeApi$migration_complete()}}
\item \href{#method-migration_status}{\code{DkubeApi$migration_status()}}
\item \href{#method-migration_status_user}{\code{DkubeApi$migration_status_user()}}
\item \href{#method-model_deploy}{\code{DkubeApi$model_deploy()}}
\item \href{#method-notification}{\code{DkubeApi$notification()}}
\item \href{#method-projects_delete_list}{\code{DkubeApi$projects_delete_list()}}
\item \href{#method-repo_update_one}{\code{DkubeApi$repo_update_one()}}
\item \href{#method-revoke_session_key}{\code{DkubeApi$revoke_session_key()}}
\item \href{#method-run_group_update}{\code{DkubeApi$run_group_update()}}
\item \href{#method-source_migration_init}{\code{DkubeApi$source_migration_init()}}
\item \href{#method-start_tensorboard}{\code{DkubeApi$start_tensorboard()}}
\item \href{#method-stop_tensorboard}{\code{DkubeApi$stop_tensorboard()}}
\item \href{#method-storage_volume_export}{\code{DkubeApi$storage_volume_export()}}
\item \href{#method-storage_volume_reclaim}{\code{DkubeApi$storage_volume_reclaim()}}
\item \href{#method-submit_project_metrics}{\code{DkubeApi$submit_project_metrics()}}
\item \href{#method-target_migration_init}{\code{DkubeApi$target_migration_init()}}
\item \href{#method-tokeninfo}{\code{DkubeApi$tokeninfo()}}
\item \href{#method-trigger_runs_by_condition}{\code{DkubeApi$trigger_runs_by_condition()}}
\item \href{#method-update_datum_version_model}{\code{DkubeApi$update_datum_version_model()}}
\item \href{#method-update_inference}{\code{DkubeApi$update_inference()}}
\item \href{#method-update_model_catalog}{\code{DkubeApi$update_model_catalog()}}
\item \href{#method-update_one_project}{\code{DkubeApi$update_one_project()}}
\item \href{#method-update_run_one}{\code{DkubeApi$update_run_one()}}
\item \href{#method-update_template_one}{\code{DkubeApi$update_template_one()}}
\item \href{#method-user_capability}{\code{DkubeApi$user_capability()}}
\item \href{#method-user_get_collection_one}{\code{DkubeApi$user_get_collection_one()}}
\item \href{#method-users_get_collection_all}{\code{DkubeApi$users_get_collection_all()}}
\item \href{#method-volumes_storage}{\code{DkubeApi$volumes_storage()}}
\item \href{#method-clone}{\code{DkubeApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-abort_migration"></a>}}
\if{latex}{\out{\hypertarget{method-abort_migration}{}}}
\subsection{Method \code{abort_migration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$abort_migration(migration, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_cicd_image"></a>}}
\if{latex}{\out{\hypertarget{method-add_cicd_image}{}}}
\subsection{Method \code{add_cicd_image()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$add_cicd_image(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-attach_user_profile"></a>}}
\if{latex}{\out{\hypertarget{method-attach_user_profile}{}}}
\subsection{Method \code{attach_user_profile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$attach_user_profile(profile, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_project"></a>}}
\if{latex}{\out{\hypertarget{method-create_project}{}}}
\subsection{Method \code{create_project()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$create_project(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_user_profile"></a>}}
\if{latex}{\out{\hypertarget{method-create_user_profile}{}}}
\subsection{Method \code{create_user_profile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$create_user_profile(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-datum_one_abort_download"></a>}}
\if{latex}{\out{\hypertarget{method-datum_one_abort_download}{}}}
\subsection{Method \code{datum_one_abort_download()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$datum_one_abort_download(user, class, datum, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-datum_one_restart_download"></a>}}
\if{latex}{\out{\hypertarget{method-datum_one_restart_download}{}}}
\subsection{Method \code{datum_one_restart_download()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$datum_one_restart_download(user, class, datum, data, extract, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-datum_refresh_one"></a>}}
\if{latex}{\out{\hypertarget{method-datum_refresh_one}{}}}
\subsection{Method \code{datum_refresh_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$datum_refresh_one(user, class, datum, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-datums_add_one"></a>}}
\if{latex}{\out{\hypertarget{method-datums_add_one}{}}}
\subsection{Method \code{datums_add_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$datums_add_one(user, data, extract, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-datums_archive_by_class"></a>}}
\if{latex}{\out{\hypertarget{method-datums_archive_by_class}{}}}
\subsection{Method \code{datums_archive_by_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$datums_archive_by_class(user, class, archive, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-datums_delete_by_class"></a>}}
\if{latex}{\out{\hypertarget{method-datums_delete_by_class}{}}}
\subsection{Method \code{datums_delete_by_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$datums_delete_by_class(user, class, data, force, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-datums_deploy_one_model"></a>}}
\if{latex}{\out{\hypertarget{method-datums_deploy_one_model}{}}}
\subsection{Method \code{datums_deploy_one_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$datums_deploy_one_model(user, model, version, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-datums_get_by_class"></a>}}
\if{latex}{\out{\hypertarget{method-datums_get_by_class}{}}}
\subsection{Method \code{datums_get_by_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$datums_get_by_class(
  user,
  class,
  shared,
  archived,
  tags,
  page,
  versions,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-datums_get_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-datums_get_metrics}{}}}
\subsection{Method \code{datums_get_metrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$datums_get_metrics(user, class, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-datums_get_one_by_class"></a>}}
\if{latex}{\out{\hypertarget{method-datums_get_one_by_class}{}}}
\subsection{Method \code{datums_get_one_by_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$datums_get_one_by_class(user, class, datum, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-datums_get_one_version_lineage"></a>}}
\if{latex}{\out{\hypertarget{method-datums_get_one_version_lineage}{}}}
\subsection{Method \code{datums_get_one_version_lineage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$datums_get_one_version_lineage(user, class, datum, version, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-datums_get_one_version_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-datums_get_one_version_metrics}{}}}
\subsection{Method \code{datums_get_one_version_metrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$datums_get_one_version_metrics(user, class, datum, version, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-datums_publish_one_model"></a>}}
\if{latex}{\out{\hypertarget{method-datums_publish_one_model}{}}}
\subsection{Method \code{datums_publish_one_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$datums_publish_one_model(user, model, version, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-datums_release_one_model"></a>}}
\if{latex}{\out{\hypertarget{method-datums_release_one_model}{}}}
\subsection{Method \code{datums_release_one_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$datums_release_one_model(user, model, version, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-datums_testdeploy_one_model"></a>}}
\if{latex}{\out{\hypertarget{method-datums_testdeploy_one_model}{}}}
\subsection{Method \code{datums_testdeploy_one_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$datums_testdeploy_one_model(user, model, version, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_cicd_images"></a>}}
\if{latex}{\out{\hypertarget{method-delete_cicd_images}{}}}
\subsection{Method \code{delete_cicd_images()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$delete_cicd_images(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_migration"></a>}}
\if{latex}{\out{\hypertarget{method-delete_migration}{}}}
\subsection{Method \code{delete_migration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$delete_migration(migration, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_model_catalog_item"></a>}}
\if{latex}{\out{\hypertarget{method-delete_model_catalog_item}{}}}
\subsection{Method \code{delete_model_catalog_item()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$delete_model_catalog_item(user, model, version, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_pipeline_experiment"></a>}}
\if{latex}{\out{\hypertarget{method-delete_pipeline_experiment}{}}}
\subsection{Method \code{delete_pipeline_experiment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$delete_pipeline_experiment(name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_pipeline_experimentby_id"></a>}}
\if{latex}{\out{\hypertarget{method-delete_pipeline_experimentby_id}{}}}
\subsection{Method \code{delete_pipeline_experimentby_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$delete_pipeline_experimentby_id(expid, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_pipeline_experiments_all"></a>}}
\if{latex}{\out{\hypertarget{method-delete_pipeline_experiments_all}{}}}
\subsection{Method \code{delete_pipeline_experiments_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$delete_pipeline_experiments_all(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_pipeline_run"></a>}}
\if{latex}{\out{\hypertarget{method-delete_pipeline_run}{}}}
\subsection{Method \code{delete_pipeline_run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$delete_pipeline_run(name, exp, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_pipeline_run_by_id"></a>}}
\if{latex}{\out{\hypertarget{method-delete_pipeline_run_by_id}{}}}
\subsection{Method \code{delete_pipeline_run_by_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$delete_pipeline_run_by_id(runid, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_pipeline_runs_all"></a>}}
\if{latex}{\out{\hypertarget{method-delete_pipeline_runs_all}{}}}
\subsection{Method \code{delete_pipeline_runs_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$delete_pipeline_runs_all(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-diag"></a>}}
\if{latex}{\out{\hypertarget{method-diag}{}}}
\subsection{Method \code{diag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$diag(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dkubeinfo"></a>}}
\if{latex}{\out{\hypertarget{method-dkubeinfo}{}}}
\subsection{Method \code{dkubeinfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$dkubeinfo(authmode, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dl_frameworks"></a>}}
\if{latex}{\out{\hypertarget{method-dl_frameworks}{}}}
\subsection{Method \code{dl_frameworks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$dl_frameworks(class, project, user, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dlsupport"></a>}}
\if{latex}{\out{\hypertarget{method-dlsupport}{}}}
\subsection{Method \code{dlsupport()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$dlsupport(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureset_add_one"></a>}}
\if{latex}{\out{\hypertarget{method-featureset_add_one}{}}}
\subsection{Method \code{featureset_add_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$featureset_add_one(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureset_commit_version"></a>}}
\if{latex}{\out{\hypertarget{method-featureset_commit_version}{}}}
\subsection{Method \code{featureset_commit_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$featureset_commit_version(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureset_copy_version"></a>}}
\if{latex}{\out{\hypertarget{method-featureset_copy_version}{}}}
\subsection{Method \code{featureset_copy_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$featureset_copy_version(featureset, version, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureset_copy_version_abort"></a>}}
\if{latex}{\out{\hypertarget{method-featureset_copy_version_abort}{}}}
\subsection{Method \code{featureset_copy_version_abort()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$featureset_copy_version_abort(featureset, version, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureset_copy_version_status"></a>}}
\if{latex}{\out{\hypertarget{method-featureset_copy_version_status}{}}}
\subsection{Method \code{featureset_copy_version_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$featureset_copy_version_status(featureset, version, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureset_delete"></a>}}
\if{latex}{\out{\hypertarget{method-featureset_delete}{}}}
\subsection{Method \code{featureset_delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$featureset_delete(data, force, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureset_get"></a>}}
\if{latex}{\out{\hypertarget{method-featureset_get}{}}}
\subsection{Method \code{featureset_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$featureset_get(featureset, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureset_get_one_version_lineage"></a>}}
\if{latex}{\out{\hypertarget{method-featureset_get_one_version_lineage}{}}}
\subsection{Method \code{featureset_get_one_version_lineage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$featureset_get_one_version_lineage(featureset, version, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureset_list"></a>}}
\if{latex}{\out{\hypertarget{method-featureset_list}{}}}
\subsection{Method \code{featureset_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$featureset_list(query, versions, tags, archived, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureset_put"></a>}}
\if{latex}{\out{\hypertarget{method-featureset_put}{}}}
\subsection{Method \code{featureset_put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$featureset_put(featureset, data, indexing, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureset_update_version"></a>}}
\if{latex}{\out{\hypertarget{method-featureset_update_version}{}}}
\subsection{Method \code{featureset_update_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$featureset_update_version(featureset, version, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featuresets_archive"></a>}}
\if{latex}{\out{\hypertarget{method-featuresets_archive}{}}}
\subsection{Method \code{featuresets_archive()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$featuresets_archive(archive, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featurespec_list"></a>}}
\if{latex}{\out{\hypertarget{method-featurespec_list}{}}}
\subsection{Method \code{featurespec_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$featurespec_list(featureset, query, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-g_p_us_capability"></a>}}
\if{latex}{\out{\hypertarget{method-g_p_us_capability}{}}}
\subsection{Method \code{g_p_us_capability()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$g_p_us_capability(source, value, class, user, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_all_cicd_images"></a>}}
\if{latex}{\out{\hypertarget{method-get_all_cicd_images}{}}}
\subsection{Method \code{get_all_cicd_images()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$get_all_cicd_images(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_all_datum_prev_cred"></a>}}
\if{latex}{\out{\hypertarget{method-get_all_datum_prev_cred}{}}}
\subsection{Method \code{get_all_datum_prev_cred()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$get_all_datum_prev_cred(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_all_project_submissions"></a>}}
\if{latex}{\out{\hypertarget{method-get_all_project_submissions}{}}}
\subsection{Method \code{get_all_project_submissions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$get_all_project_submissions(project_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_all_projects"></a>}}
\if{latex}{\out{\hypertarget{method-get_all_projects}{}}}
\subsection{Method \code{get_all_projects()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$get_all_projects(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_datum_prev_cred"></a>}}
\if{latex}{\out{\hypertarget{method-get_datum_prev_cred}{}}}
\subsection{Method \code{get_datum_prev_cred()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$get_datum_prev_cred(user, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dvs_config"></a>}}
\if{latex}{\out{\hypertarget{method-get_dvs_config}{}}}
\subsection{Method \code{get_dvs_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$get_dvs_config(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_job_groups"></a>}}
\if{latex}{\out{\hypertarget{method-get_job_groups}{}}}
\subsection{Method \code{get_job_groups()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$get_job_groups(user, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_migration_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-get_migration_metadata}{}}}
\subsection{Method \code{get_migration_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$get_migration_metadata(migration, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_one_project"></a>}}
\if{latex}{\out{\hypertarget{method-get_one_project}{}}}
\subsection{Method \code{get_one_project()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$get_one_project(project_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_one_run_lineage"></a>}}
\if{latex}{\out{\hypertarget{method-get_one_run_lineage}{}}}
\subsection{Method \code{get_one_run_lineage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$get_one_run_lineage(user, class, runid, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_role"></a>}}
\if{latex}{\out{\hypertarget{method-get_role}{}}}
\subsection{Method \code{get_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$get_role(rolename, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_session_key"></a>}}
\if{latex}{\out{\hypertarget{method-get_session_key}{}}}
\subsection{Method \code{get_session_key()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$get_session_key(user, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_single_user_profile"></a>}}
\if{latex}{\out{\hypertarget{method-get_single_user_profile}{}}}
\subsection{Method \code{get_single_user_profile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$get_single_user_profile(profile, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inference_monitoring"></a>}}
\if{latex}{\out{\hypertarget{method-inference_monitoring}{}}}
\subsection{Method \code{inference_monitoring()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$inference_monitoring(user, job, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-job_get_collection_all_by_class"></a>}}
\if{latex}{\out{\hypertarget{method-job_get_collection_all_by_class}{}}}
\subsection{Method \code{job_get_collection_all_by_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$job_get_collection_all_by_class(
  user,
  class,
  run,
  template,
  group,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-job_group_add_one"></a>}}
\if{latex}{\out{\hypertarget{method-job_group_add_one}{}}}
\subsection{Method \code{job_group_add_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$job_group_add_one(user, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jobmigration"></a>}}
\if{latex}{\out{\hypertarget{method-jobmigration}{}}}
\subsection{Method \code{jobmigration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$jobmigration(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jobs_abort_one"></a>}}
\if{latex}{\out{\hypertarget{method-jobs_abort_one}{}}}
\subsection{Method \code{jobs_abort_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$jobs_abort_one(user, job, class, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jobs_add_one"></a>}}
\if{latex}{\out{\hypertarget{method-jobs_add_one}{}}}
\subsection{Method \code{jobs_add_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$jobs_add_one(user, data, run, rerun, subclass, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jobs_delete_one_by_uuid"></a>}}
\if{latex}{\out{\hypertarget{method-jobs_delete_one_by_uuid}{}}}
\subsection{Method \code{jobs_delete_one_by_uuid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$jobs_delete_one_by_uuid(uuid, class, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jobs_get_by_class"></a>}}
\if{latex}{\out{\hypertarget{method-jobs_get_by_class}{}}}
\subsection{Method \code{jobs_get_by_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$jobs_get_by_class(
  user,
  class,
  shared,
  run,
  template,
  group,
  all,
  archived,
  tags,
  page,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jobs_get_collection_all"></a>}}
\if{latex}{\out{\hypertarget{method-jobs_get_collection_all}{}}}
\subsection{Method \code{jobs_get_collection_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$jobs_get_collection_all(user, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jobs_get_collection_one"></a>}}
\if{latex}{\out{\hypertarget{method-jobs_get_collection_one}{}}}
\subsection{Method \code{jobs_get_collection_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$jobs_get_collection_one(user, class, job, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jobs_get_one_by_uuid"></a>}}
\if{latex}{\out{\hypertarget{method-jobs_get_one_by_uuid}{}}}
\subsection{Method \code{jobs_get_one_by_uuid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$jobs_get_one_by_uuid(uuid, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jobs_list_archive_by_class"></a>}}
\if{latex}{\out{\hypertarget{method-jobs_list_archive_by_class}{}}}
\subsection{Method \code{jobs_list_archive_by_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$jobs_list_archive_by_class(user, class, archive, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jobs_list_delete_by_class"></a>}}
\if{latex}{\out{\hypertarget{method-jobs_list_delete_by_class}{}}}
\subsection{Method \code{jobs_list_delete_by_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$jobs_list_delete_by_class(user, class, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jobs_start_one"></a>}}
\if{latex}{\out{\hypertarget{method-jobs_start_one}{}}}
\subsection{Method \code{jobs_start_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$jobs_start_one(user, job, class, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jsonform"></a>}}
\if{latex}{\out{\hypertarget{method-jsonform}{}}}
\subsection{Method \code{jsonform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$jsonform(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_user_profiles"></a>}}
\if{latex}{\out{\hypertarget{method-list_user_profiles}{}}}
\subsection{Method \code{list_user_profiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$list_user_profiles(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-login"></a>}}
\if{latex}{\out{\hypertarget{method-login}{}}}
\subsection{Method \code{login()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$login(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logout"></a>}}
\if{latex}{\out{\hypertarget{method-logout}{}}}
\subsection{Method \code{logout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$logout(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logs_get_one"></a>}}
\if{latex}{\out{\hypertarget{method-logs_get_one}{}}}
\subsection{Method \code{logs_get_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$logs_get_one(user, job, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-migration_complete"></a>}}
\if{latex}{\out{\hypertarget{method-migration_complete}{}}}
\subsection{Method \code{migration_complete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$migration_complete(migration, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-migration_status"></a>}}
\if{latex}{\out{\hypertarget{method-migration_status}{}}}
\subsection{Method \code{migration_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$migration_status(migration, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-migration_status_user"></a>}}
\if{latex}{\out{\hypertarget{method-migration_status_user}{}}}
\subsection{Method \code{migration_status_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$migration_status_user(migration, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model_deploy"></a>}}
\if{latex}{\out{\hypertarget{method-model_deploy}{}}}
\subsection{Method \code{model_deploy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$model_deploy(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-notification"></a>}}
\if{latex}{\out{\hypertarget{method-notification}{}}}
\subsection{Method \code{notification()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$notification(type, op, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-projects_delete_list"></a>}}
\if{latex}{\out{\hypertarget{method-projects_delete_list}{}}}
\subsection{Method \code{projects_delete_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$projects_delete_list(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repo_update_one"></a>}}
\if{latex}{\out{\hypertarget{method-repo_update_one}{}}}
\subsection{Method \code{repo_update_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$repo_update_one(user, class, repo, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-revoke_session_key"></a>}}
\if{latex}{\out{\hypertarget{method-revoke_session_key}{}}}
\subsection{Method \code{revoke_session_key()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$revoke_session_key(user, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_group_update"></a>}}
\if{latex}{\out{\hypertarget{method-run_group_update}{}}}
\subsection{Method \code{run_group_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$run_group_update(user, rungroup, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-source_migration_init"></a>}}
\if{latex}{\out{\hypertarget{method-source_migration_init}{}}}
\subsection{Method \code{source_migration_init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$source_migration_init(migration, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_tensorboard"></a>}}
\if{latex}{\out{\hypertarget{method-start_tensorboard}{}}}
\subsection{Method \code{start_tensorboard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$start_tensorboard(user, class, job, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop_tensorboard"></a>}}
\if{latex}{\out{\hypertarget{method-stop_tensorboard}{}}}
\subsection{Method \code{stop_tensorboard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$stop_tensorboard(user, class, job, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-storage_volume_export"></a>}}
\if{latex}{\out{\hypertarget{method-storage_volume_export}{}}}
\subsection{Method \code{storage_volume_export()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$storage_volume_export(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-storage_volume_reclaim"></a>}}
\if{latex}{\out{\hypertarget{method-storage_volume_reclaim}{}}}
\subsection{Method \code{storage_volume_reclaim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$storage_volume_reclaim(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submit_project_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-submit_project_metrics}{}}}
\subsection{Method \code{submit_project_metrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$submit_project_metrics(project_id, submission_id, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-target_migration_init"></a>}}
\if{latex}{\out{\hypertarget{method-target_migration_init}{}}}
\subsection{Method \code{target_migration_init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$target_migration_init(migration, upfile, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tokeninfo"></a>}}
\if{latex}{\out{\hypertarget{method-tokeninfo}{}}}
\subsection{Method \code{tokeninfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$tokeninfo(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trigger_runs_by_condition"></a>}}
\if{latex}{\out{\hypertarget{method-trigger_runs_by_condition}{}}}
\subsection{Method \code{trigger_runs_by_condition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$trigger_runs_by_condition(conditions, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_datum_version_model"></a>}}
\if{latex}{\out{\hypertarget{method-update_datum_version_model}{}}}
\subsection{Method \code{update_datum_version_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$update_datum_version_model(user, uuid, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_inference"></a>}}
\if{latex}{\out{\hypertarget{method-update_inference}{}}}
\subsection{Method \code{update_inference()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$update_inference(user, job, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_model_catalog"></a>}}
\if{latex}{\out{\hypertarget{method-update_model_catalog}{}}}
\subsection{Method \code{update_model_catalog()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$update_model_catalog(user, model, version, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_one_project"></a>}}
\if{latex}{\out{\hypertarget{method-update_one_project}{}}}
\subsection{Method \code{update_one_project()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$update_one_project(project_id, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_run_one"></a>}}
\if{latex}{\out{\hypertarget{method-update_run_one}{}}}
\subsection{Method \code{update_run_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$update_run_one(user, class, run, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_template_one"></a>}}
\if{latex}{\out{\hypertarget{method-update_template_one}{}}}
\subsection{Method \code{update_template_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$update_template_one(user, class, template, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_capability"></a>}}
\if{latex}{\out{\hypertarget{method-user_capability}{}}}
\subsection{Method \code{user_capability()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$user_capability(user, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_get_collection_one"></a>}}
\if{latex}{\out{\hypertarget{method-user_get_collection_one}{}}}
\subsection{Method \code{user_get_collection_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$user_get_collection_one(user, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-users_get_collection_all"></a>}}
\if{latex}{\out{\hypertarget{method-users_get_collection_all}{}}}
\subsection{Method \code{users_get_collection_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$users_get_collection_all(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-volumes_storage"></a>}}
\if{latex}{\out{\hypertarget{method-volumes_storage}{}}}
\subsection{Method \code{volumes_storage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$volumes_storage(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
