% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DkubeContainerModelFrameworkDetails.r
\name{DkubeContainerModelFrameworkDetails}
\alias{DkubeContainerModelFrameworkDetails}
\title{DkubeContainerModelFrameworkDetails Class}
\description{
DkubeContainerModelFrameworkDetails Class

DkubeContainerModelFrameworkDetails Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{version}}{}

\item{\code{image}}{}

\item{\code{isprivate}}{}

\item{\code{username}}{}

\item{\code{password}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DkubeContainerModelFrameworkDetails$new()}}
\item \href{#method-toJSON}{\code{DkubeContainerModelFrameworkDetails$toJSON()}}
\item \href{#method-fromJSON}{\code{DkubeContainerModelFrameworkDetails$fromJSON()}}
\item \href{#method-toJSONString}{\code{DkubeContainerModelFrameworkDetails$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DkubeContainerModelFrameworkDetails$fromJSONString()}}
\item \href{#method-clone}{\code{DkubeContainerModelFrameworkDetails$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeContainerModelFrameworkDetails$new(
  version,
  image,
  isprivate,
  username,
  password
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeContainerModelFrameworkDetails$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeContainerModelFrameworkDetails$fromJSON(
  DkubeContainerModelFrameworkDetailsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeContainerModelFrameworkDetails$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeContainerModelFrameworkDetails$fromJSONString(
  DkubeContainerModelFrameworkDetailsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeContainerModelFrameworkDetails$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
