% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DkubeInfoLicense.r
\name{DkubeInfoLicense}
\alias{DkubeInfoLicense}
\title{DkubeInfoLicense Class}
\description{
DkubeInfoLicense Class

DkubeInfoLicense Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{startDate}}{}

\item{\code{endDate}}{}

\item{\code{validity}}{}

\item{\code{licensee}}{}

\item{\code{genTime}}{}

\item{\code{expired}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DkubeInfoLicense$new()}}
\item \href{#method-toJSON}{\code{DkubeInfoLicense$toJSON()}}
\item \href{#method-fromJSON}{\code{DkubeInfoLicense$fromJSON()}}
\item \href{#method-toJSONString}{\code{DkubeInfoLicense$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DkubeInfoLicense$fromJSONString()}}
\item \href{#method-clone}{\code{DkubeInfoLicense$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeInfoLicense$new(startDate, endDate, validity, licensee, genTime, expired)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeInfoLicense$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeInfoLicense$fromJSON(DkubeInfoLicenseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeInfoLicense$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeInfoLicense$fromJSONString(DkubeInfoLicenseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeInfoLicense$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
