% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DkubeOperatorExclusiveApi.r
\name{DkubeOperatorExclusiveApi}
\alias{DkubeOperatorExclusiveApi}
\title{DkubeOperatorExclusive operations}
\description{
dkube_api.DkubeOperatorExclusive
}
\section{Methods}{

\describe{

add_dvs_config API to configuring dkube version management system.


add_role API to add a role.


change_auth API to change the current authorization mode.


configure_clusters API to configure execution clusters to dkube.


delete_cluster API to delete a cluster configuration from dkube


delete_dvs_config API to delete dkube version management system configuration.


featurestore_get API to get featurestore configuration


featurestore_put API to edit faturestore configuration


get_all_roles API to get all roles.


get_auth API to fetch information about all authorization modes.


get_cluster_details API to get the details for a cluster


get_clusters API to fetch configured clusters


get_model_catalog API to fetch the model catalog


get_model_catalog_name API to fetch the model catalog name for the given model


get_released_models API to fetch the released model


init API to initialize the dkube. The user initializing should be an admin and will be auto onboarded into dkube with operator role.


license_update API to update dkube license.


log_viewer_delete API to remove log viewer configuration from Fluentd


log_viewer_get API to get log viewer configuration


log_viewer_post API to enable Fluentd to talk to a log viewer


log_viewer_put API to edit log viewer configuration in Fluentd


modelcatalog_get_metrics API to get metrics of a list of modelcatalogs


offboard API to offboard a user into dkube.


onboard API to onboard a user into dkube.


plugin_create_one API to create a plugin.


plugin_upgrade_one API to upgrade a deployed plugin.


plugins_list_all API to fetch list of all bundled and external plugins


plugins_list_delete_by_name API to delete plugins by name


recover API to recover the D3 system.


role_update API to promote/demote a user.


sync API to sync D3 db with the resources available in cluster This also synchronizes members from github.


sync_group_users API to sync D3 db with the resources available in cluster This also synchronizes members from github.


update_cluster API to update a cluster configuration.


update_model_in_catalog API to update a model in model catalog


update_role API to update a role.

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DkubeOperatorExclusiveApi$new()}}
\item \href{#method-add_dvs_config}{\code{DkubeOperatorExclusiveApi$add_dvs_config()}}
\item \href{#method-add_role}{\code{DkubeOperatorExclusiveApi$add_role()}}
\item \href{#method-change_auth}{\code{DkubeOperatorExclusiveApi$change_auth()}}
\item \href{#method-configure_clusters}{\code{DkubeOperatorExclusiveApi$configure_clusters()}}
\item \href{#method-delete_cluster}{\code{DkubeOperatorExclusiveApi$delete_cluster()}}
\item \href{#method-delete_dvs_config}{\code{DkubeOperatorExclusiveApi$delete_dvs_config()}}
\item \href{#method-featurestore_get}{\code{DkubeOperatorExclusiveApi$featurestore_get()}}
\item \href{#method-featurestore_put}{\code{DkubeOperatorExclusiveApi$featurestore_put()}}
\item \href{#method-get_all_roles}{\code{DkubeOperatorExclusiveApi$get_all_roles()}}
\item \href{#method-get_auth}{\code{DkubeOperatorExclusiveApi$get_auth()}}
\item \href{#method-get_cluster_details}{\code{DkubeOperatorExclusiveApi$get_cluster_details()}}
\item \href{#method-get_clusters}{\code{DkubeOperatorExclusiveApi$get_clusters()}}
\item \href{#method-get_model_catalog}{\code{DkubeOperatorExclusiveApi$get_model_catalog()}}
\item \href{#method-get_model_catalog_name}{\code{DkubeOperatorExclusiveApi$get_model_catalog_name()}}
\item \href{#method-get_released_models}{\code{DkubeOperatorExclusiveApi$get_released_models()}}
\item \href{#method-init}{\code{DkubeOperatorExclusiveApi$init()}}
\item \href{#method-license_update}{\code{DkubeOperatorExclusiveApi$license_update()}}
\item \href{#method-log_viewer_delete}{\code{DkubeOperatorExclusiveApi$log_viewer_delete()}}
\item \href{#method-log_viewer_get}{\code{DkubeOperatorExclusiveApi$log_viewer_get()}}
\item \href{#method-log_viewer_post}{\code{DkubeOperatorExclusiveApi$log_viewer_post()}}
\item \href{#method-log_viewer_put}{\code{DkubeOperatorExclusiveApi$log_viewer_put()}}
\item \href{#method-modelcatalog_get_metrics}{\code{DkubeOperatorExclusiveApi$modelcatalog_get_metrics()}}
\item \href{#method-offboard}{\code{DkubeOperatorExclusiveApi$offboard()}}
\item \href{#method-onboard}{\code{DkubeOperatorExclusiveApi$onboard()}}
\item \href{#method-plugin_create_one}{\code{DkubeOperatorExclusiveApi$plugin_create_one()}}
\item \href{#method-plugin_upgrade_one}{\code{DkubeOperatorExclusiveApi$plugin_upgrade_one()}}
\item \href{#method-plugins_list_all}{\code{DkubeOperatorExclusiveApi$plugins_list_all()}}
\item \href{#method-plugins_list_delete_by_name}{\code{DkubeOperatorExclusiveApi$plugins_list_delete_by_name()}}
\item \href{#method-recover}{\code{DkubeOperatorExclusiveApi$recover()}}
\item \href{#method-role_update}{\code{DkubeOperatorExclusiveApi$role_update()}}
\item \href{#method-sync}{\code{DkubeOperatorExclusiveApi$sync()}}
\item \href{#method-sync_group_users}{\code{DkubeOperatorExclusiveApi$sync_group_users()}}
\item \href{#method-update_cluster}{\code{DkubeOperatorExclusiveApi$update_cluster()}}
\item \href{#method-update_model_in_catalog}{\code{DkubeOperatorExclusiveApi$update_model_in_catalog()}}
\item \href{#method-update_role}{\code{DkubeOperatorExclusiveApi$update_role()}}
\item \href{#method-clone}{\code{DkubeOperatorExclusiveApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_dvs_config"></a>}}
\if{latex}{\out{\hypertarget{method-add_dvs_config}{}}}
\subsection{Method \code{add_dvs_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$add_dvs_config(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_role"></a>}}
\if{latex}{\out{\hypertarget{method-add_role}{}}}
\subsection{Method \code{add_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$add_role(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-change_auth"></a>}}
\if{latex}{\out{\hypertarget{method-change_auth}{}}}
\subsection{Method \code{change_auth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$change_auth(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-configure_clusters"></a>}}
\if{latex}{\out{\hypertarget{method-configure_clusters}{}}}
\subsection{Method \code{configure_clusters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$configure_clusters(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_cluster"></a>}}
\if{latex}{\out{\hypertarget{method-delete_cluster}{}}}
\subsection{Method \code{delete_cluster()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$delete_cluster(clustername, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_dvs_config"></a>}}
\if{latex}{\out{\hypertarget{method-delete_dvs_config}{}}}
\subsection{Method \code{delete_dvs_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$delete_dvs_config(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featurestore_get"></a>}}
\if{latex}{\out{\hypertarget{method-featurestore_get}{}}}
\subsection{Method \code{featurestore_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$featurestore_get(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featurestore_put"></a>}}
\if{latex}{\out{\hypertarget{method-featurestore_put}{}}}
\subsection{Method \code{featurestore_put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$featurestore_put(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_all_roles"></a>}}
\if{latex}{\out{\hypertarget{method-get_all_roles}{}}}
\subsection{Method \code{get_all_roles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$get_all_roles(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_auth"></a>}}
\if{latex}{\out{\hypertarget{method-get_auth}{}}}
\subsection{Method \code{get_auth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$get_auth(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_cluster_details"></a>}}
\if{latex}{\out{\hypertarget{method-get_cluster_details}{}}}
\subsection{Method \code{get_cluster_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$get_cluster_details(clustername, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_clusters"></a>}}
\if{latex}{\out{\hypertarget{method-get_clusters}{}}}
\subsection{Method \code{get_clusters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$get_clusters(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_model_catalog"></a>}}
\if{latex}{\out{\hypertarget{method-get_model_catalog}{}}}
\subsection{Method \code{get_model_catalog()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$get_model_catalog(user, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_model_catalog_name"></a>}}
\if{latex}{\out{\hypertarget{method-get_model_catalog_name}{}}}
\subsection{Method \code{get_model_catalog_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$get_model_catalog_name(user, model, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_released_models"></a>}}
\if{latex}{\out{\hypertarget{method-get_released_models}{}}}
\subsection{Method \code{get_released_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$get_released_models(user, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$init(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-license_update"></a>}}
\if{latex}{\out{\hypertarget{method-license_update}{}}}
\subsection{Method \code{license_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$license_update(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_viewer_delete"></a>}}
\if{latex}{\out{\hypertarget{method-log_viewer_delete}{}}}
\subsection{Method \code{log_viewer_delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$log_viewer_delete(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_viewer_get"></a>}}
\if{latex}{\out{\hypertarget{method-log_viewer_get}{}}}
\subsection{Method \code{log_viewer_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$log_viewer_get(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_viewer_post"></a>}}
\if{latex}{\out{\hypertarget{method-log_viewer_post}{}}}
\subsection{Method \code{log_viewer_post()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$log_viewer_post(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_viewer_put"></a>}}
\if{latex}{\out{\hypertarget{method-log_viewer_put}{}}}
\subsection{Method \code{log_viewer_put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$log_viewer_put(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-modelcatalog_get_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-modelcatalog_get_metrics}{}}}
\subsection{Method \code{modelcatalog_get_metrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$modelcatalog_get_metrics(user, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-offboard"></a>}}
\if{latex}{\out{\hypertarget{method-offboard}{}}}
\subsection{Method \code{offboard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$offboard(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-onboard"></a>}}
\if{latex}{\out{\hypertarget{method-onboard}{}}}
\subsection{Method \code{onboard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$onboard(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plugin_create_one"></a>}}
\if{latex}{\out{\hypertarget{method-plugin_create_one}{}}}
\subsection{Method \code{plugin_create_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$plugin_create_one(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plugin_upgrade_one"></a>}}
\if{latex}{\out{\hypertarget{method-plugin_upgrade_one}{}}}
\subsection{Method \code{plugin_upgrade_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$plugin_upgrade_one(plugin, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plugins_list_all"></a>}}
\if{latex}{\out{\hypertarget{method-plugins_list_all}{}}}
\subsection{Method \code{plugins_list_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$plugins_list_all(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plugins_list_delete_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-plugins_list_delete_by_name}{}}}
\subsection{Method \code{plugins_list_delete_by_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$plugins_list_delete_by_name(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recover"></a>}}
\if{latex}{\out{\hypertarget{method-recover}{}}}
\subsection{Method \code{recover()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$recover(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-role_update"></a>}}
\if{latex}{\out{\hypertarget{method-role_update}{}}}
\subsection{Method \code{role_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$role_update(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sync"></a>}}
\if{latex}{\out{\hypertarget{method-sync}{}}}
\subsection{Method \code{sync()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$sync(kind, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sync_group_users"></a>}}
\if{latex}{\out{\hypertarget{method-sync_group_users}{}}}
\subsection{Method \code{sync_group_users()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$sync_group_users(group, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_cluster"></a>}}
\if{latex}{\out{\hypertarget{method-update_cluster}{}}}
\subsection{Method \code{update_cluster()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$update_cluster(clustername, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_model_in_catalog"></a>}}
\if{latex}{\out{\hypertarget{method-update_model_in_catalog}{}}}
\subsection{Method \code{update_model_in_catalog()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$update_model_in_catalog(
  user,
  model,
  version,
  data,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_role"></a>}}
\if{latex}{\out{\hypertarget{method-update_role}{}}}
\subsection{Method \code{update_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$update_role(rolename, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DkubeOperatorExclusiveApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
