% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLFrameworks.r
\name{DLFrameworks}
\alias{DLFrameworks}
\title{DLFrameworks Class}
\description{
DLFrameworks Class

DLFrameworks Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{nb-ide}}{}

\item{\code{r-ide}}{}

\item{\code{training}}{}

\item{\code{preprocessing}}{}

\item{\code{serving}}{}

\item{\code{project}}{}

\item{\code{frameworks}}{}

\item{\code{images}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DLFrameworks$new()}}
\item \href{#method-toJSON}{\code{DLFrameworks$toJSON()}}
\item \href{#method-fromJSON}{\code{DLFrameworks$fromJSON()}}
\item \href{#method-toJSONString}{\code{DLFrameworks$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DLFrameworks$fromJSONString()}}
\item \href{#method-clone}{\code{DLFrameworks$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DLFrameworks$new(
  `nb-ide`,
  `r-ide`,
  training,
  preprocessing,
  serving,
  project,
  frameworks,
  images
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DLFrameworks$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DLFrameworks$fromJSON(DLFrameworksJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DLFrameworks$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DLFrameworks$fromJSONString(DLFrameworksJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DLFrameworks$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
