% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSJobModel.r
\name{DSJobModel}
\alias{DSJobModel}
\title{DSJobModel Class}
\description{
DSJobModel Class

DSJobModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{executor}}{}

\item{\code{datums}}{}

\item{\code{featuresets}}{}

\item{\code{tags}}{}

\item{\code{hptuning}}{}

\item{\code{hyperparams}}{}

\item{\code{nworkers}}{}

\item{\code{ngpus}}{}

\item{\code{rdma}}{}

\item{\code{gpus_override}}{}

\item{\code{view}}{}

\item{\code{cluster}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DSJobModel$new()}}
\item \href{#method-toJSON}{\code{DSJobModel$toJSON()}}
\item \href{#method-fromJSON}{\code{DSJobModel$fromJSON()}}
\item \href{#method-toJSONString}{\code{DSJobModel$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DSJobModel$fromJSONString()}}
\item \href{#method-clone}{\code{DSJobModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DSJobModel$new(
  executor,
  datums,
  featuresets,
  tags,
  hptuning,
  hyperparams,
  nworkers,
  ngpus,
  rdma,
  gpus_override,
  view,
  cluster
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DSJobModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DSJobModel$fromJSON(DSJobModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DSJobModel$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DSJobModel$fromJSONString(DSJobModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DSJobModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
