% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSJobModelHyperparams.r
\name{DSJobModelHyperparams}
\alias{DSJobModelHyperparams}
\title{DSJobModelHyperparams Class}
\description{
DSJobModelHyperparams Class

DSJobModelHyperparams Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{epochs}}{}

\item{\code{steps}}{}

\item{\code{batchsize}}{}

\item{\code{custom}}{}

\item{\code{customkv}}{}

\item{\code{file}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DSJobModelHyperparams$new()}}
\item \href{#method-toJSON}{\code{DSJobModelHyperparams$toJSON()}}
\item \href{#method-fromJSON}{\code{DSJobModelHyperparams$fromJSON()}}
\item \href{#method-toJSONString}{\code{DSJobModelHyperparams$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DSJobModelHyperparams$fromJSONString()}}
\item \href{#method-clone}{\code{DSJobModelHyperparams$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DSJobModelHyperparams$new(epochs, steps, batchsize, custom, customkv, file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DSJobModelHyperparams$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DSJobModelHyperparams$fromJSON(DSJobModelHyperparamsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DSJobModelHyperparams$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DSJobModelHyperparams$fromJSONString(DSJobModelHyperparamsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DSJobModelHyperparams$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
