% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSetDef.r
\name{FeatureSetDef}
\alias{FeatureSetDef}
\title{FeatureSetDef Class}
\description{
FeatureSetDef Class

FeatureSetDef Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{tags}}{}

\item{\code{owner}}{}

\item{\code{created_ts}}{}

\item{\code{updated_ts}}{}

\item{\code{job}}{}

\item{\code{status}}{}

\item{\code{uuid}}{}

\item{\code{deleted}}{}

\item{\code{archived}}{}

\item{\code{originalName}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FeatureSetDef$new()}}
\item \href{#method-toJSON}{\code{FeatureSetDef$toJSON()}}
\item \href{#method-fromJSON}{\code{FeatureSetDef$fromJSON()}}
\item \href{#method-toJSONString}{\code{FeatureSetDef$toJSONString()}}
\item \href{#method-fromJSONString}{\code{FeatureSetDef$fromJSONString()}}
\item \href{#method-clone}{\code{FeatureSetDef$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetDef$new(
  name,
  description,
  tags,
  owner,
  created_ts,
  updated_ts,
  job,
  status,
  uuid,
  deleted,
  archived,
  originalName
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetDef$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetDef$fromJSON(FeatureSetDefJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetDef$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetDef$fromJSONString(FeatureSetDefJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetDef$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
