% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesetLineageInOutModel.r
\name{FeaturesetLineageInOutModel}
\alias{FeaturesetLineageInOutModel}
\title{FeaturesetLineageInOutModel Class}
\description{
FeaturesetLineageInOutModel Class

FeaturesetLineageInOutModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{}

\item{\code{uuid}}{}

\item{\code{version_name}}{}

\item{\code{version_uuid}}{}

\item{\code{type}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FeaturesetLineageInOutModel$new()}}
\item \href{#method-toJSON}{\code{FeaturesetLineageInOutModel$toJSON()}}
\item \href{#method-fromJSON}{\code{FeaturesetLineageInOutModel$fromJSON()}}
\item \href{#method-toJSONString}{\code{FeaturesetLineageInOutModel$toJSONString()}}
\item \href{#method-fromJSONString}{\code{FeaturesetLineageInOutModel$fromJSONString()}}
\item \href{#method-clone}{\code{FeaturesetLineageInOutModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeaturesetLineageInOutModel$new(name, uuid, version_name, version_uuid, type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeaturesetLineageInOutModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeaturesetLineageInOutModel$fromJSON(FeaturesetLineageInOutModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeaturesetLineageInOutModel$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeaturesetLineageInOutModel$fromJSONString(FeaturesetLineageInOutModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeaturesetLineageInOutModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
