% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSetUsageDef.r
\name{FeatureSetUsageDef}
\alias{FeatureSetUsageDef}
\title{FeatureSetUsageDef Class}
\description{
FeatureSetUsageDef Class

FeatureSetUsageDef Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{}

\item{\code{user}}{}

\item{\code{jobid}}{}

\item{\code{class}}{}

\item{\code{state}}{}

\item{\code{tags}}{}

\item{\code{pipeline}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FeatureSetUsageDef$new()}}
\item \href{#method-toJSON}{\code{FeatureSetUsageDef$toJSON()}}
\item \href{#method-fromJSON}{\code{FeatureSetUsageDef$fromJSON()}}
\item \href{#method-toJSONString}{\code{FeatureSetUsageDef$toJSONString()}}
\item \href{#method-fromJSONString}{\code{FeatureSetUsageDef$fromJSONString()}}
\item \href{#method-clone}{\code{FeatureSetUsageDef$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetUsageDef$new(name, user, jobid, class, state, tags, pipeline)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetUsageDef$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetUsageDef$fromJSON(FeatureSetUsageDefJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetUsageDef$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetUsageDef$fromJSONString(FeatureSetUsageDefJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetUsageDef$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
