% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSetVersionDef.r
\name{FeatureSetVersionDef}
\alias{FeatureSetVersionDef}
\title{FeatureSetVersionDef Class}
\description{
FeatureSetVersionDef Class

FeatureSetVersionDef Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{}

\item{\code{uuid}}{}

\item{\code{tags}}{}

\item{\code{description}}{}

\item{\code{index}}{}

\item{\code{created_ts}}{}

\item{\code{updated_ts}}{}

\item{\code{state}}{}

\item{\code{user}}{}

\item{\code{featuresetName}}{}

\item{\code{featuresetUUID}}{}

\item{\code{job}}{}

\item{\code{trackingUUID}}{}

\item{\code{commitID}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FeatureSetVersionDef$new()}}
\item \href{#method-toJSON}{\code{FeatureSetVersionDef$toJSON()}}
\item \href{#method-fromJSON}{\code{FeatureSetVersionDef$fromJSON()}}
\item \href{#method-toJSONString}{\code{FeatureSetVersionDef$toJSONString()}}
\item \href{#method-fromJSONString}{\code{FeatureSetVersionDef$fromJSONString()}}
\item \href{#method-clone}{\code{FeatureSetVersionDef$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetVersionDef$new(
  name,
  uuid,
  tags,
  description,
  index,
  created_ts,
  updated_ts,
  state,
  user,
  featuresetName,
  featuresetUUID,
  job,
  trackingUUID,
  commitID
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetVersionDef$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetVersionDef$fromJSON(FeatureSetVersionDefJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetVersionDef$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetVersionDef$fromJSONString(FeatureSetVersionDefJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetVersionDef$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
