% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceJobModel.r
\name{InferenceJobModel}
\alias{InferenceJobModel}
\title{InferenceJobModel Class}
\description{
InferenceJobModel Class

InferenceJobModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{}

\item{\code{version}}{}

\item{\code{owner}}{}

\item{\code{device}}{}

\item{\code{deploy}}{}

\item{\code{mcname}}{}

\item{\code{serving_image}}{}

\item{\code{transformer}}{}

\item{\code{transformer_image}}{}

\item{\code{transformer_project}}{}

\item{\code{transformer-commitID}}{}

\item{\code{transformer_code}}{}

\item{\code{minReplicas}}{}

\item{\code{maxConcurrentRequests}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InferenceJobModel$new()}}
\item \href{#method-toJSON}{\code{InferenceJobModel$toJSON()}}
\item \href{#method-fromJSON}{\code{InferenceJobModel$fromJSON()}}
\item \href{#method-toJSONString}{\code{InferenceJobModel$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InferenceJobModel$fromJSONString()}}
\item \href{#method-clone}{\code{InferenceJobModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceJobModel$new(
  model,
  version,
  owner,
  device,
  deploy,
  mcname,
  serving_image,
  transformer,
  transformer_image,
  transformer_project,
  `transformer-commitID`,
  transformer_code,
  minReplicas,
  maxConcurrentRequests
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceJobModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceJobModel$fromJSON(InferenceJobModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceJobModel$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceJobModel$fromJSONString(InferenceJobModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceJobModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
