% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceMonitoringOutModel.r
\name{InferenceMonitoringOutModel}
\alias{InferenceMonitoringOutModel}
\title{InferenceMonitoringOutModel Class}
\description{
InferenceMonitoringOutModel Class

InferenceMonitoringOutModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{job}}{}

\item{\code{prometheus_query}}{}

\item{\code{deployments}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InferenceMonitoringOutModel$new()}}
\item \href{#method-toJSON}{\code{InferenceMonitoringOutModel$toJSON()}}
\item \href{#method-fromJSON}{\code{InferenceMonitoringOutModel$fromJSON()}}
\item \href{#method-toJSONString}{\code{InferenceMonitoringOutModel$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InferenceMonitoringOutModel$fromJSONString()}}
\item \href{#method-clone}{\code{InferenceMonitoringOutModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceMonitoringOutModel$new(job, prometheus_query, deployments)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceMonitoringOutModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceMonitoringOutModel$fromJSON(InferenceMonitoringOutModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceMonitoringOutModel$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceMonitoringOutModel$fromJSONString(InferenceMonitoringOutModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceMonitoringOutModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
