% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InlineResponse20030DataOutputs.r
\name{InlineResponse20030DataOutputs}
\alias{InlineResponse20030DataOutputs}
\title{InlineResponse20030DataOutputs Class}
\description{
InlineResponse20030DataOutputs Class

InlineResponse20030DataOutputs Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datum}}{}

\item{\code{version}}{}

\item{\code{featureset}}{}

\item{\code{featureset_version}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InlineResponse20030DataOutputs$new()}}
\item \href{#method-toJSON}{\code{InlineResponse20030DataOutputs$toJSON()}}
\item \href{#method-fromJSON}{\code{InlineResponse20030DataOutputs$fromJSON()}}
\item \href{#method-toJSONString}{\code{InlineResponse20030DataOutputs$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InlineResponse20030DataOutputs$fromJSONString()}}
\item \href{#method-clone}{\code{InlineResponse20030DataOutputs$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20030DataOutputs$new(
  datum,
  version,
  featureset,
  featureset_version
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20030DataOutputs$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20030DataOutputs$fromJSON(InlineResponse20030DataOutputsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20030DataOutputs$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20030DataOutputs$fromJSONString(
  InlineResponse20030DataOutputsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse20030DataOutputs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
