% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobClusterModelSchedulingOptsSlurmhpc.r
\name{JobClusterModelSchedulingOptsSlurmhpc}
\alias{JobClusterModelSchedulingOptsSlurmhpc}
\title{JobClusterModelSchedulingOptsSlurmhpc Class}
\description{
JobClusterModelSchedulingOptsSlurmhpc Class

JobClusterModelSchedulingOptsSlurmhpc Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{file}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{JobClusterModelSchedulingOptsSlurmhpc$new()}}
\item \href{#method-toJSON}{\code{JobClusterModelSchedulingOptsSlurmhpc$toJSON()}}
\item \href{#method-fromJSON}{\code{JobClusterModelSchedulingOptsSlurmhpc$fromJSON()}}
\item \href{#method-toJSONString}{\code{JobClusterModelSchedulingOptsSlurmhpc$toJSONString()}}
\item \href{#method-fromJSONString}{\code{JobClusterModelSchedulingOptsSlurmhpc$fromJSONString()}}
\item \href{#method-clone}{\code{JobClusterModelSchedulingOptsSlurmhpc$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobClusterModelSchedulingOptsSlurmhpc$new(file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobClusterModelSchedulingOptsSlurmhpc$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobClusterModelSchedulingOptsSlurmhpc$fromJSON(
  JobClusterModelSchedulingOptsSlurmhpcJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobClusterModelSchedulingOptsSlurmhpc$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobClusterModelSchedulingOptsSlurmhpc$fromJSONString(
  JobClusterModelSchedulingOptsSlurmhpcJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobClusterModelSchedulingOptsSlurmhpc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
