% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobInputFeaturesetModel.r
\name{JobInputFeaturesetModel}
\alias{JobInputFeaturesetModel}
\title{JobInputFeaturesetModel Class}
\description{
JobInputFeaturesetModel Class

JobInputFeaturesetModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{}

\item{\code{version}}{}

\item{\code{owner}}{}

\item{\code{mountpath}}{}

\item{\code{version_name}}{}

\item{\code{job_added}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{JobInputFeaturesetModel$new()}}
\item \href{#method-toJSON}{\code{JobInputFeaturesetModel$toJSON()}}
\item \href{#method-fromJSON}{\code{JobInputFeaturesetModel$fromJSON()}}
\item \href{#method-toJSONString}{\code{JobInputFeaturesetModel$toJSONString()}}
\item \href{#method-fromJSONString}{\code{JobInputFeaturesetModel$fromJSONString()}}
\item \href{#method-clone}{\code{JobInputFeaturesetModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobInputFeaturesetModel$new(
  name,
  version,
  owner,
  mountpath,
  version_name,
  job_added
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobInputFeaturesetModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobInputFeaturesetModel$fromJSON(JobInputFeaturesetModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobInputFeaturesetModel$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobInputFeaturesetModel$fromJSONString(JobInputFeaturesetModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobInputFeaturesetModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
