% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobModelParameters.r
\name{JobModelParameters}
\alias{JobModelParameters}
\title{JobModelParameters Class}
\description{
JobModelParameters Class

JobModelParameters Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{class}}{}

\item{\code{gpu_allocation}}{}

\item{\code{priority}}{}

\item{\code{training}}{}

\item{\code{notebook}}{}

\item{\code{inference}}{}

\item{\code{preprocessing}}{}

\item{\code{custom}}{}

\item{\code{run}}{}

\item{\code{generated}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{JobModelParameters$new()}}
\item \href{#method-toJSON}{\code{JobModelParameters$toJSON()}}
\item \href{#method-fromJSON}{\code{JobModelParameters$fromJSON()}}
\item \href{#method-toJSONString}{\code{JobModelParameters$toJSONString()}}
\item \href{#method-fromJSONString}{\code{JobModelParameters$fromJSONString()}}
\item \href{#method-clone}{\code{JobModelParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParameters$new(
  class,
  gpu_allocation,
  priority,
  training,
  notebook,
  inference,
  preprocessing,
  custom,
  run,
  generated
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParameters$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParameters$fromJSON(JobModelParametersJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParameters$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParameters$fromJSONString(JobModelParametersJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
