% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobModelParametersGenerated.r
\name{JobModelParametersGenerated}
\alias{JobModelParametersGenerated}
\title{JobModelParametersGenerated Class}
\description{
JobModelParametersGenerated Class

JobModelParametersGenerated Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{deleted}}{}

\item{\code{archived}}{}

\item{\code{isprivate}}{}

\item{\code{category}}{}

\item{\code{cicdstatus}}{}

\item{\code{pipeline}}{}

\item{\code{subclass}}{}

\item{\code{hp_tuning_info}}{}

\item{\code{training_image}}{}

\item{\code{serving_image}}{}

\item{\code{scheduled}}{}

\item{\code{affinity}}{}

\item{\code{user}}{}

\item{\code{jobid}}{}

\item{\code{uuid}}{}

\item{\code{status}}{}

\item{\code{timestamps}}{}

\item{\code{runtime}}{}

\item{\code{ngpus_alloc}}{}

\item{\code{ngpus_max}}{}

\item{\code{tbref}}{}

\item{\code{studyref}}{}

\item{\code{trialref}}{}

\item{\code{best_trial_id}}{}

\item{\code{best_objective_value}}{}

\item{\code{accelerator}}{}

\item{\code{versions}}{}

\item{\code{tracking}}{}

\item{\code{inputDatumRefs}}{}

\item{\code{inputFeaturesetRefs}}{}

\item{\code{details}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{JobModelParametersGenerated$new()}}
\item \href{#method-toJSON}{\code{JobModelParametersGenerated$toJSON()}}
\item \href{#method-fromJSON}{\code{JobModelParametersGenerated$fromJSON()}}
\item \href{#method-toJSONString}{\code{JobModelParametersGenerated$toJSONString()}}
\item \href{#method-fromJSONString}{\code{JobModelParametersGenerated$fromJSONString()}}
\item \href{#method-clone}{\code{JobModelParametersGenerated$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParametersGenerated$new(
  deleted,
  archived,
  isprivate,
  category,
  cicdstatus,
  pipeline,
  subclass,
  hp_tuning_info,
  training_image,
  serving_image,
  scheduled,
  affinity,
  user,
  jobid,
  uuid,
  status,
  timestamps,
  runtime,
  ngpus_alloc,
  ngpus_max,
  tbref,
  studyref,
  trialref,
  best_trial_id,
  best_objective_value,
  accelerator,
  versions,
  tracking,
  inputDatumRefs,
  inputFeaturesetRefs,
  details
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParametersGenerated$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParametersGenerated$fromJSON(JobModelParametersGeneratedJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParametersGenerated$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParametersGenerated$fromJSONString(JobModelParametersGeneratedJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParametersGenerated$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
