% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobModelParametersGeneratedDetails.r
\name{JobModelParametersGeneratedDetails}
\alias{JobModelParametersGeneratedDetails}
\title{JobModelParametersGeneratedDetails Class}
\description{
JobModelParametersGeneratedDetails Class

JobModelParametersGeneratedDetails Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tensorboard}}{}

\item{\code{notebook}}{}

\item{\code{training}}{}

\item{\code{serving}}{}

\item{\code{custom}}{}

\item{\code{accesstoken}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{JobModelParametersGeneratedDetails$new()}}
\item \href{#method-toJSON}{\code{JobModelParametersGeneratedDetails$toJSON()}}
\item \href{#method-fromJSON}{\code{JobModelParametersGeneratedDetails$fromJSON()}}
\item \href{#method-toJSONString}{\code{JobModelParametersGeneratedDetails$toJSONString()}}
\item \href{#method-fromJSONString}{\code{JobModelParametersGeneratedDetails$fromJSONString()}}
\item \href{#method-clone}{\code{JobModelParametersGeneratedDetails$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParametersGeneratedDetails$new(
  tensorboard,
  notebook,
  training,
  serving,
  custom,
  accesstoken
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParametersGeneratedDetails$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParametersGeneratedDetails$fromJSON(
  JobModelParametersGeneratedDetailsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParametersGeneratedDetails$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParametersGeneratedDetails$fromJSONString(
  JobModelParametersGeneratedDetailsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParametersGeneratedDetails$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
