% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobModelParametersGeneratedInputDatumRefs.r
\name{JobModelParametersGeneratedInputDatumRefs}
\alias{JobModelParametersGeneratedInputDatumRefs}
\title{JobModelParametersGeneratedInputDatumRefs Class}
\description{
JobModelParametersGeneratedInputDatumRefs Class

JobModelParametersGeneratedInputDatumRefs Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{class}}{}

\item{\code{uuid}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{JobModelParametersGeneratedInputDatumRefs$new()}}
\item \href{#method-toJSON}{\code{JobModelParametersGeneratedInputDatumRefs$toJSON()}}
\item \href{#method-fromJSON}{\code{JobModelParametersGeneratedInputDatumRefs$fromJSON()}}
\item \href{#method-toJSONString}{\code{JobModelParametersGeneratedInputDatumRefs$toJSONString()}}
\item \href{#method-fromJSONString}{\code{JobModelParametersGeneratedInputDatumRefs$fromJSONString()}}
\item \href{#method-clone}{\code{JobModelParametersGeneratedInputDatumRefs$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParametersGeneratedInputDatumRefs$new(class, uuid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParametersGeneratedInputDatumRefs$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParametersGeneratedInputDatumRefs$fromJSON(
  JobModelParametersGeneratedInputDatumRefsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParametersGeneratedInputDatumRefs$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParametersGeneratedInputDatumRefs$fromJSONString(
  JobModelParametersGeneratedInputDatumRefsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobModelParametersGeneratedInputDatumRefs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
