% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MigrationDatumModel.r
\name{MigrationDatumModel}
\alias{MigrationDatumModel}
\title{MigrationDatumModel Class}
\description{
MigrationDatumModel Class

MigrationDatumModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{UUID}}{}

\item{\code{remote_UUID}}{}

\item{\code{dkubeDatumUUIDs}}{}

\item{\code{migrationUUID}}{}

\item{\code{migrationJobUUID}}{}

\item{\code{duplicate}}{}

\item{\code{duplicate_ref}}{}

\item{\code{datum_type}}{}

\item{\code{bucket}}{}

\item{\code{status}}{}

\item{\code{progress}}{}

\item{\code{details}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MigrationDatumModel$new()}}
\item \href{#method-toJSON}{\code{MigrationDatumModel$toJSON()}}
\item \href{#method-fromJSON}{\code{MigrationDatumModel$fromJSON()}}
\item \href{#method-toJSONString}{\code{MigrationDatumModel$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MigrationDatumModel$fromJSONString()}}
\item \href{#method-clone}{\code{MigrationDatumModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationDatumModel$new(
  UUID,
  remote_UUID,
  dkubeDatumUUIDs,
  migrationUUID,
  migrationJobUUID,
  duplicate,
  duplicate_ref,
  datum_type,
  bucket,
  status,
  progress,
  details
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationDatumModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationDatumModel$fromJSON(MigrationDatumModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationDatumModel$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationDatumModel$fromJSONString(MigrationDatumModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationDatumModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
