% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MigrationModel.r
\name{MigrationModel}
\alias{MigrationModel}
\title{MigrationModel Class}
\description{
MigrationModel Class

MigrationModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{version}}{}

\item{\code{name}}{}

\item{\code{UUID}}{}

\item{\code{remote_UUID}}{}

\item{\code{User}}{}

\item{\code{triggering_point}}{}

\item{\code{location}}{}

\item{\code{remote}}{}

\item{\code{progress}}{}

\item{\code{status}}{}

\item{\code{time_stamp}}{}

\item{\code{jobs}}{}

\item{\code{job_refs}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MigrationModel$new()}}
\item \href{#method-toJSON}{\code{MigrationModel$toJSON()}}
\item \href{#method-fromJSON}{\code{MigrationModel$fromJSON()}}
\item \href{#method-toJSONString}{\code{MigrationModel$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MigrationModel$fromJSONString()}}
\item \href{#method-clone}{\code{MigrationModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationModel$new(
  version,
  name,
  UUID,
  remote_UUID,
  User,
  triggering_point,
  location,
  remote,
  progress,
  status,
  time_stamp,
  jobs,
  job_refs
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationModel$fromJSON(MigrationModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationModel$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationModel$fromJSONString(MigrationModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
