% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelDetailsKindDkubeTrained.r
\name{ModelDetailsKindDkubeTrained}
\alias{ModelDetailsKindDkubeTrained}
\title{ModelDetailsKindDkubeTrained Class}
\description{
ModelDetailsKindDkubeTrained Class

ModelDetailsKindDkubeTrained Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{job}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelDetailsKindDkubeTrained$new()}}
\item \href{#method-toJSON}{\code{ModelDetailsKindDkubeTrained$toJSON()}}
\item \href{#method-fromJSON}{\code{ModelDetailsKindDkubeTrained$fromJSON()}}
\item \href{#method-toJSONString}{\code{ModelDetailsKindDkubeTrained$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ModelDetailsKindDkubeTrained$fromJSONString()}}
\item \href{#method-clone}{\code{ModelDetailsKindDkubeTrained$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDetailsKindDkubeTrained$new(job)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDetailsKindDkubeTrained$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDetailsKindDkubeTrained$fromJSON(ModelDetailsKindDkubeTrainedJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDetailsKindDkubeTrained$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDetailsKindDkubeTrained$fromJSONString(ModelDetailsKindDkubeTrainedJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDetailsKindDkubeTrained$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
