% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelDetailsTensorpb.r
\name{ModelDetailsTensorpb}
\alias{ModelDetailsTensorpb}
\title{ModelDetailsTensorpb Class}
\description{
ModelDetailsTensorpb Class

ModelDetailsTensorpb Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{signatures}}{}

\item{\code{devicenodes}}{}

\item{\code{devices}}{}

\item{\code{parameters}}{}

\item{\code{weights}}{}

\item{\code{layers}}{}

\item{\code{servable}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelDetailsTensorpb$new()}}
\item \href{#method-toJSON}{\code{ModelDetailsTensorpb$toJSON()}}
\item \href{#method-fromJSON}{\code{ModelDetailsTensorpb$fromJSON()}}
\item \href{#method-toJSONString}{\code{ModelDetailsTensorpb$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ModelDetailsTensorpb$fromJSONString()}}
\item \href{#method-clone}{\code{ModelDetailsTensorpb$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDetailsTensorpb$new(
  signatures,
  devicenodes,
  devices,
  parameters,
  weights,
  layers,
  servable
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDetailsTensorpb$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDetailsTensorpb$fromJSON(ModelDetailsTensorpbJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDetailsTensorpb$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDetailsTensorpb$fromJSONString(ModelDetailsTensorpbJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDetailsTensorpb$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
