% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelDetailsTensorpbDevices.r
\name{ModelDetailsTensorpbDevices}
\alias{ModelDetailsTensorpbDevices}
\title{ModelDetailsTensorpbDevices Class}
\description{
ModelDetailsTensorpbDevices Class

ModelDetailsTensorpbDevices Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cpu}}{}

\item{\code{gpu}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelDetailsTensorpbDevices$new()}}
\item \href{#method-toJSON}{\code{ModelDetailsTensorpbDevices$toJSON()}}
\item \href{#method-fromJSON}{\code{ModelDetailsTensorpbDevices$fromJSON()}}
\item \href{#method-toJSONString}{\code{ModelDetailsTensorpbDevices$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ModelDetailsTensorpbDevices$fromJSONString()}}
\item \href{#method-clone}{\code{ModelDetailsTensorpbDevices$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDetailsTensorpbDevices$new(cpu, gpu)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDetailsTensorpbDevices$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDetailsTensorpbDevices$fromJSON(ModelDetailsTensorpbDevicesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDetailsTensorpbDevices$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDetailsTensorpbDevices$fromJSONString(ModelDetailsTensorpbDevicesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDetailsTensorpbDevices$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
