% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelTrainingInfoDatasetsSourceDetails.r
\name{ModelTrainingInfoDatasetsSourceDetails}
\alias{ModelTrainingInfoDatasetsSourceDetails}
\title{ModelTrainingInfoDatasetsSourceDetails Class}
\description{
ModelTrainingInfoDatasetsSourceDetails Class

ModelTrainingInfoDatasetsSourceDetails Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{source}}{}

\item{\code{version}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelTrainingInfoDatasetsSourceDetails$new()}}
\item \href{#method-toJSON}{\code{ModelTrainingInfoDatasetsSourceDetails$toJSON()}}
\item \href{#method-fromJSON}{\code{ModelTrainingInfoDatasetsSourceDetails$fromJSON()}}
\item \href{#method-toJSONString}{\code{ModelTrainingInfoDatasetsSourceDetails$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ModelTrainingInfoDatasetsSourceDetails$fromJSONString()}}
\item \href{#method-clone}{\code{ModelTrainingInfoDatasetsSourceDetails$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelTrainingInfoDatasetsSourceDetails$new(source, version)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelTrainingInfoDatasetsSourceDetails$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelTrainingInfoDatasetsSourceDetails$fromJSON(
  ModelTrainingInfoDatasetsSourceDetailsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelTrainingInfoDatasetsSourceDetails$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelTrainingInfoDatasetsSourceDetails$fromJSONString(
  ModelTrainingInfoDatasetsSourceDetailsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelTrainingInfoDatasetsSourceDetails$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
