% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelTrainingInfoFeaturesets.r
\name{ModelTrainingInfoFeaturesets}
\alias{ModelTrainingInfoFeaturesets}
\title{ModelTrainingInfoFeaturesets Class}
\description{
ModelTrainingInfoFeaturesets Class

ModelTrainingInfoFeaturesets Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{}

\item{\code{version}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelTrainingInfoFeaturesets$new()}}
\item \href{#method-toJSON}{\code{ModelTrainingInfoFeaturesets$toJSON()}}
\item \href{#method-fromJSON}{\code{ModelTrainingInfoFeaturesets$fromJSON()}}
\item \href{#method-toJSONString}{\code{ModelTrainingInfoFeaturesets$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ModelTrainingInfoFeaturesets$fromJSONString()}}
\item \href{#method-clone}{\code{ModelTrainingInfoFeaturesets$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelTrainingInfoFeaturesets$new(name, version)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelTrainingInfoFeaturesets$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelTrainingInfoFeaturesets$fromJSON(ModelTrainingInfoFeaturesetsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelTrainingInfoFeaturesets$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelTrainingInfoFeaturesets$fromJSONString(ModelTrainingInfoFeaturesetsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelTrainingInfoFeaturesets$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
