% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NotificationData.r
\name{NotificationData}
\alias{NotificationData}
\title{NotificationData Class}
\description{
NotificationData Class

NotificationData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Username}}{}

\item{\code{JobUUID}}{}

\item{\code{Class}}{}

\item{\code{Subclass}}{}

\item{\code{Status}}{}

\item{\code{Reason}}{}

\item{\code{NotebookUrl}}{}

\item{\code{ServingUrl}}{}

\item{\code{TensorboardUrl}}{}

\item{\code{IsTensorboard}}{}

\item{\code{Study}}{}

\item{\code{Worker}}{}

\item{\code{Datum}}{}

\item{\code{Version}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NotificationData$new()}}
\item \href{#method-toJSON}{\code{NotificationData$toJSON()}}
\item \href{#method-fromJSON}{\code{NotificationData$fromJSON()}}
\item \href{#method-toJSONString}{\code{NotificationData$toJSONString()}}
\item \href{#method-fromJSONString}{\code{NotificationData$fromJSONString()}}
\item \href{#method-clone}{\code{NotificationData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationData$new(
  Username,
  JobUUID,
  Class,
  Subclass,
  Status,
  Reason,
  NotebookUrl,
  ServingUrl,
  TensorboardUrl,
  IsTensorboard,
  Study,
  Worker,
  Datum,
  Version
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationData$fromJSON(NotificationDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationData$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationData$fromJSONString(NotificationDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
