% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessingJobModel.r
\name{PreprocessingJobModel}
\alias{PreprocessingJobModel}
\title{PreprocessingJobModel Class}
\description{
PreprocessingJobModel Class

PreprocessingJobModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{kind}}{}

\item{\code{executor}}{}

\item{\code{datums}}{}

\item{\code{featuresets}}{}

\item{\code{tags}}{}

\item{\code{config}}{}

\item{\code{cluster}}{}

\item{\code{ngpus}}{}

\item{\code{gpus_override}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PreprocessingJobModel$new()}}
\item \href{#method-toJSON}{\code{PreprocessingJobModel$toJSON()}}
\item \href{#method-fromJSON}{\code{PreprocessingJobModel$fromJSON()}}
\item \href{#method-toJSONString}{\code{PreprocessingJobModel$toJSONString()}}
\item \href{#method-fromJSONString}{\code{PreprocessingJobModel$fromJSONString()}}
\item \href{#method-clone}{\code{PreprocessingJobModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreprocessingJobModel$new(
  kind,
  executor,
  datums,
  featuresets,
  tags,
  config,
  cluster,
  ngpus,
  gpus_override
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreprocessingJobModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreprocessingJobModel$fromJSON(PreprocessingJobModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreprocessingJobModel$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreprocessingJobModel$fromJSONString(PreprocessingJobModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreprocessingJobModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
