% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessingJobModelExecutor.r
\name{PreprocessingJobModelExecutor}
\alias{PreprocessingJobModelExecutor}
\title{PreprocessingJobModelExecutor Class}
\description{
PreprocessingJobModelExecutor Class

PreprocessingJobModelExecutor Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{choice}}{}

\item{\code{buildfromproject}}{}

\item{\code{custom}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PreprocessingJobModelExecutor$new()}}
\item \href{#method-toJSON}{\code{PreprocessingJobModelExecutor$toJSON()}}
\item \href{#method-fromJSON}{\code{PreprocessingJobModelExecutor$fromJSON()}}
\item \href{#method-toJSONString}{\code{PreprocessingJobModelExecutor$toJSONString()}}
\item \href{#method-fromJSONString}{\code{PreprocessingJobModelExecutor$fromJSONString()}}
\item \href{#method-clone}{\code{PreprocessingJobModelExecutor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreprocessingJobModelExecutor$new(choice, buildfromproject, custom)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreprocessingJobModelExecutor$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreprocessingJobModelExecutor$fromJSON(PreprocessingJobModelExecutorJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreprocessingJobModelExecutor$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreprocessingJobModelExecutor$fromJSONString(PreprocessingJobModelExecutorJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreprocessingJobModelExecutor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
