% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectUpdateModel.r
\name{ProjectUpdateModel}
\alias{ProjectUpdateModel}
\title{ProjectUpdateModel Class}
\description{
ProjectUpdateModel Class

ProjectUpdateModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name}

\item{\code{description}}{Description}

\item{\code{image}}{Image}

\item{\code{leaderboard}}{Leaderboard}

\item{\code{details}}{Project Details}

\item{\code{eval_repo}}{Eval Repo}

\item{\code{eval_commit_id}}{Eval Commit Id}

\item{\code{eval_image}}{Eval Image}

\item{\code{eval_script}}{Eval Script}

\item{\code{eval_details}}{Eval Details}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProjectUpdateModel$new()}}
\item \href{#method-toJSON}{\code{ProjectUpdateModel$toJSON()}}
\item \href{#method-fromJSON}{\code{ProjectUpdateModel$fromJSON()}}
\item \href{#method-toJSONString}{\code{ProjectUpdateModel$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ProjectUpdateModel$fromJSONString()}}
\item \href{#method-clone}{\code{ProjectUpdateModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectUpdateModel$new(
  name,
  description,
  image,
  leaderboard,
  details,
  eval_repo,
  eval_commit_id,
  eval_image,
  eval_script,
  eval_details
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectUpdateModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectUpdateModel$fromJSON(ProjectUpdateModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectUpdateModel$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectUpdateModel$fromJSONString(ProjectUpdateModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectUpdateModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
