% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedshiftAccessInfo.r
\name{RedshiftAccessInfo}
\alias{RedshiftAccessInfo}
\title{RedshiftAccessInfo Class}
\description{
RedshiftAccessInfo Class

RedshiftAccessInfo Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{endpoint}}{}

\item{\code{username}}{}

\item{\code{password}}{}

\item{\code{database}}{}

\item{\code{region}}{}

\item{\code{cacert}}{}

\item{\code{insecureSSL}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RedshiftAccessInfo$new()}}
\item \href{#method-toJSON}{\code{RedshiftAccessInfo$toJSON()}}
\item \href{#method-fromJSON}{\code{RedshiftAccessInfo$fromJSON()}}
\item \href{#method-toJSONString}{\code{RedshiftAccessInfo$toJSONString()}}
\item \href{#method-fromJSONString}{\code{RedshiftAccessInfo$fromJSONString()}}
\item \href{#method-clone}{\code{RedshiftAccessInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedshiftAccessInfo$new(
  endpoint,
  username,
  password,
  database,
  region,
  cacert,
  insecureSSL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedshiftAccessInfo$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedshiftAccessInfo$fromJSON(RedshiftAccessInfoJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedshiftAccessInfo$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedshiftAccessInfo$fromJSONString(RedshiftAccessInfoJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedshiftAccessInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
