% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReleasedModelItem.r
\name{ReleasedModelItem}
\alias{ReleasedModelItem}
\title{ReleasedModelItem Class}
\description{
ReleasedModelItem Class

ReleasedModelItem Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{}

\item{\code{reason}}{}

\item{\code{version}}{}

\item{\code{user}}{}

\item{\code{stage}}{}

\item{\code{trackingURL}}{}

\item{\code{UUID}}{}

\item{\code{description}}{}

\item{\code{releasedTimestamp}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ReleasedModelItem$new()}}
\item \href{#method-toJSON}{\code{ReleasedModelItem$toJSON()}}
\item \href{#method-fromJSON}{\code{ReleasedModelItem$fromJSON()}}
\item \href{#method-toJSONString}{\code{ReleasedModelItem$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ReleasedModelItem$fromJSONString()}}
\item \href{#method-clone}{\code{ReleasedModelItem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReleasedModelItem$new(
  name,
  reason,
  version,
  user,
  stage,
  trackingURL,
  UUID,
  description,
  releasedTimestamp
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReleasedModelItem$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReleasedModelItem$fromJSON(ReleasedModelItemJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReleasedModelItem$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReleasedModelItem$fromJSONString(ReleasedModelItemJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReleasedModelItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
