% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunTemplateModelParameters.r
\name{RunTemplateModelParameters}
\alias{RunTemplateModelParameters}
\title{RunTemplateModelParameters Class}
\description{
RunTemplateModelParameters Class

RunTemplateModelParameters Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gpu_allocation}}{}

\item{\code{priority}}{}

\item{\code{training}}{}

\item{\code{preprocessing}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RunTemplateModelParameters$new()}}
\item \href{#method-toJSON}{\code{RunTemplateModelParameters$toJSON()}}
\item \href{#method-fromJSON}{\code{RunTemplateModelParameters$fromJSON()}}
\item \href{#method-toJSONString}{\code{RunTemplateModelParameters$toJSONString()}}
\item \href{#method-fromJSONString}{\code{RunTemplateModelParameters$fromJSONString()}}
\item \href{#method-clone}{\code{RunTemplateModelParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunTemplateModelParameters$new(
  gpu_allocation,
  priority,
  training,
  preprocessing
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunTemplateModelParameters$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunTemplateModelParameters$fromJSON(RunTemplateModelParametersJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunTemplateModelParameters$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunTemplateModelParameters$fromJSONString(RunTemplateModelParametersJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunTemplateModelParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
