% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubmissionModel.r
\name{SubmissionModel}
\alias{SubmissionModel}
\title{SubmissionModel Class}
\description{
SubmissionModel Class

SubmissionModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Submission Id}

\item{\code{project_id}}{Project Id}

\item{\code{group_name}}{Groupname}

\item{\code{user}}{Username}

\item{\code{eval_run}}{Eval Run}

\item{\code{eval_run_name}}{Eval Run Name}

\item{\code{metrics}}{Metrics}

\item{\code{artifacts_url}}{Artifacts URL}

\item{\code{timestamp}}{Timestamp}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SubmissionModel$new()}}
\item \href{#method-toJSON}{\code{SubmissionModel$toJSON()}}
\item \href{#method-fromJSON}{\code{SubmissionModel$fromJSON()}}
\item \href{#method-toJSONString}{\code{SubmissionModel$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SubmissionModel$fromJSONString()}}
\item \href{#method-clone}{\code{SubmissionModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubmissionModel$new(
  id,
  project_id,
  group_name,
  user,
  eval_run,
  eval_run_name,
  metrics,
  artifacts_url,
  timestamp
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubmissionModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubmissionModel$fromJSON(SubmissionModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubmissionModel$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubmissionModel$fromJSONString(SubmissionModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubmissionModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
