% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VersionModel.r
\name{VersionModel}
\alias{VersionModel}
\title{VersionModel Class}
\description{
VersionModel Class

VersionModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{}

\item{\code{uuid}}{}

\item{\code{tags}}{}

\item{\code{description}}{}

\item{\code{index}}{}

\item{\code{timestamps}}{}

\item{\code{syncState}}{}

\item{\code{storagePath}}{}

\item{\code{user}}{}

\item{\code{datumName}}{}

\item{\code{datumUUID}}{}

\item{\code{runUUID}}{}

\item{\code{trackingUUID}}{}

\item{\code{commitID}}{}

\item{\code{datumType}}{}

\item{\code{model}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VersionModel$new()}}
\item \href{#method-toJSON}{\code{VersionModel$toJSON()}}
\item \href{#method-fromJSON}{\code{VersionModel$fromJSON()}}
\item \href{#method-toJSONString}{\code{VersionModel$toJSONString()}}
\item \href{#method-fromJSONString}{\code{VersionModel$fromJSONString()}}
\item \href{#method-clone}{\code{VersionModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionModel$new(
  name,
  uuid,
  tags,
  description,
  index,
  timestamps,
  syncState,
  storagePath,
  user,
  datumName,
  datumUUID,
  runUUID,
  trackingUUID,
  commitID,
  datumType,
  model
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionModel$fromJSON(VersionModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionModel$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionModel$fromJSONString(VersionModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
